% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSizeComps.R
\name{aggregateSizeComps}
\alias{aggregateSizeComps}
\title{Aggregate size comps from survey data over factors}
\usage{
aggregateSizeComps(
  zcs,
  facs = "",
  dropLevels = NULL,
  var = c("ABUNDANCE", "BIOMASS"),
  export = FALSE,
  out.csv = paste("SurveyACD", var[1], "csv", sep = "."),
  out.dir = NULL,
  verbosity = 0
)
}
\arguments{
\item{zcs}{dataframe from call to one of the calcSizeComps... functions}

\item{facs}{factors to include in output}

\item{dropLevels}{list by factor of factor levels to drop}

\item{var}{variable type to export (abundance or biomass)}

\item{export}{boolean flag to write results to csv file}

\item{out.csv}{output file name}

\item{out.dir}{output file directory}

\item{verbosity}{integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
a dataframe
}
\description{
Function to aggregate size comps from survey data over factors.
}
\details{
The time series for each factor level combination is plotted
separately. Distinct levels of each factor can be dropped from the
final plot by seting dropLevels to a list with names corresponding to
factor columns and values being vectors of factor levels to drop.
}
