% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfrf.AssignNMFSStations.R
\name{bsfrf.AssignNMFSStations}
\alias{bsfrf.AssignNMFSStations}
\title{Re-assign NMFS stations to BSFRF hauls using GIS}
\usage{
bsfrf.AssignNMFSStations(
  tbl,
  colLon = "midtowlongitude",
  colLat = "midtowlatitude"
)
}
\arguments{
\item{tbl}{: path to BSFRF csv file to read or dataframe with BSFRF haul info derived
from a dataframe created by \code{\link[=bsfrf.ReadCSV]{bsfrf.ReadCSV()}} (see @details)}

\item{colLon}{\itemize{
\item name of column with longitudes (default = "midtowlongitude")
}}

\item{colLat}{\itemize{
\item name of column with latitudes (default = "midtowlatitude")
}}
}
\value{
dataframe identical to input except that "nmfs_stn" has been reassigned
based on the NMFS survey grid cell ID in which the mid-tow location fell.
}
\description{
Function to re-assign NMFS stations to BSFRF hauls using the mid-tow location of the latter.
}
\details{
If \code{tbl} is a character string, it is assumed to be a file path; the
associated csv fle is read using \code{\link[=bsfrf.ReadCSV]{bsfrf.ReadCSV()}}. This function creates an
\pkg{sf} points layer from the resulting (or input) dataframe
using the midtowlongitude and midtowlatitude coordinates for each haul, and
determines the NMFS EBS shelf survey station that the haul midpoint falls within using
\code{\link[sf:st_join]{sf::st_join()}} with \code{join=sf::st_within} and the NMFS EBS shelf survey station grid
from \code{\link[=gisGetSurveyGridLayers]{gisGetSurveyGridLayers()}}. The NMFS station ID is copied into
(or assigned to) the column "nmfs_stn".

If \code{tbl} is not a BSFRF survey dataset filename, it must at least have columns with
lat and lon coordinates.
}
\seealso{
\code{\link[sf:geos_binary_pred]{sf::st_within()}}.
}
