% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfrf.CheckStations.R
\name{bsfrf.CheckStations}
\alias{bsfrf.CheckStations}
\title{Check BSFRF data for bad station ids}
\usage{
bsfrf.CheckStations(tbl)
}
\arguments{
\item{tbl}{: path to BSFRF csv file to read or dataframe created by \code{\link[=bsfrf.ReadCSV]{bsfrf.ReadCSV()}}}
}
\value{
dataframe with "bad" station ids
}
\description{
Function to check BSFRF data for bad station ids.
}
\details{
If \code{tbl} is a character string, it is assumed to be a file path; the
associated csv fle is read using \code{\link[=bsfrf.ReadCSV]{bsfrf.ReadCSV()}}. This function extracts the
distinct hauls in the input file/dataframe, creates an \pkg{sf} points layer
using the midtowlongitude and midtowlatitude coordinates for each haul, and
determines the NMFS grid station that the haul midpoint falls within using
\code{\link[sf:st_join]{sf::st_join()}} with \code{join=sf::st_within}.
}
\seealso{
\code{\link[sf:geos_binary_pred]{sf::st_within()}}.
}
