% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfrf.ConvertFormat2NMFS.R
\name{bsfrf.ConvertFormat2NMFS}
\alias{bsfrf.ConvertFormat2NMFS}
\title{Convert BSFRF data to (almost) NMFS CRABHAUL_DATA format}
\usage{
bsfrf.ConvertFormat2NMFS(
  tbl,
  types = c("SBS", "IDX"),
  sizeCol = "fixed_size",
  bsfrf_species = "BRD",
  nmfs_species_code = 68560,
  nmfs_species_name = "Bairdi Tanner Crab",
  verbosity = TRUE
)
}
\arguments{
\item{tbl}{: filename for BSFRF  csv file to read or dataframe from reading csv file with \code{\link[=bsfrf.ReadCSV]{bsfrf.ReadCSV()}}}

\item{types}{: BSFRF study types to select ("SBS", "IDX", or both)}

\item{bsfrf_species}{: BSFRF species to extract (default = "BRD")}

\item{nmfs_species_code}{: appropriate NMFS species code (default = 68560 for Bairdi)}

\item{nmfs_species_name}{: appropriate NMFS species name (default = 'Bairdi Tanner Crab')}

\item{verbosity}{: integer flag to print debug info}

\item{size_col}{: column name with crab sizes (default = 'fixed_cw')}
}
\value{
dataframe "almost" in CRABHAUL_DATA format.
}
\description{
Function to convert BSFRF data to (almost) NMFS CRABHAUL_DATA format.
}
\details{
BSFRF station ids are not necessarily correct. For example, the
so-called corner stations in the NMFS survey grid use the format 'aaxxxx' whereas
the BSFRF format is typically 'aa-xxxx' , where aa is a 2-letter code and
xxxx is a 4-digit code. Also, NMFS station 'H-21' has also been mis-coded 'H21' in
the BSFRF data in some years. Rather than deal with these instances piecemeal,
the NMFS survey grid layers (\code{\link[=gisGetSurveyGridLayers]{gisGetSurveyGridLayers()}}) are used to reassign NMFS station ids to the
BSFRF data based on the mid-tow coordinates for the BSFRF data. The hauls
that will be changed can be identified by running \code{\link[=bsfrf.CheckStation]{bsfrf.CheckStation()}} on the
BSFRF haul dataframe/csv file before making the conversion here.

returned dataframe has columns
\itemize{
\item {AKFIN_SURVEY_YEAR - year}
\item {HAULJOIN - paste(toupper(study),boat,start_date,start_hour,tow,sep=";")}
\item {VESSEL - boat}
\item {CRUISE - -1}
\item {HAUL - tow}
\item {HAUL_TYPE - toupper(study)}
\item {START_DATE - character: mmddyyyy}
\item {AKFIN_SURVEY_DATE - character: dd-mm-yyyy}
\item {START_HOUR - character: hhmm}
\item {MID_LATITUDE - midtowlatitude (numeric)}
\item {MID_LONGITUDE - midtowlongitude (numeric)}
\item {GIS_STATION - nmfs_stn}
\item {BOTTOM_DEPTH -  depth_ftm, converted to meters}
\item {GEAR_TEMPERATURE -  temp_c (in deg C)}
\item {AREA_SWEPT_VARIABLE -  aswept_nm2 (in nm^2)}
\item {SPECIES_CODE - nmfs_species_code}
\item {SPECIES_NAME - nmfs_species_name}
\item {SEX - sex}
\item {WIDTH - sizeCol (in mm)}
\item {SHELL_CONDITION - shell/Shell}
\item {EGG_COLOR - col}
\item {EGG_CONDITION - con}
\item {CLUTCH_SIZE - full}
\item {CHELA_HEIGHT - NA}
\item {CALCULATED_WEIGHT - NA}
\item {WEIGHT - NA}
\item {SAMPLING_FACTOR - sampfactor}
\item {CPUE_NUM - cpuenum (expanded number/nm^2)}
}
}
