% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAB.ByStratum.R
\name{calcAB.ByStratum}
\alias{calcAB.ByStratum}
\title{Calculate abundance, and biomass by stratum from AFSC trawl survey data.}
\usage{
calcAB.ByStratum(
  tbl_strata,
  tbl_cpue = NULL,
  useStratumArea = TRUE,
  export = FALSE,
  out.csv = "SurveyBiomass.ByStratum.csv",
  out.dir = NULL,
  verbosity = 0
)
}
\arguments{
\item{tbl_strata}{: data frame w/ stations/strata from call to selectStrata.TrawlSurvey(...)}

\item{tbl_cpue}{: data frame w/ cpue by year, station, other factor levels (or csv filename or NULL)}

\item{useStratumArea}{: flag (T/F) to use STRATUM_AREA to expand average CPUE to stratum abundance/biomass (default is T)}

\item{export}{: boolean flag to write results to csv file}

\item{out.csv}{: output file name}

\item{out.dir}{: output file directory}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
data frame with average cpue (numbers, weight), abundance and biomass by stratum. Columns are \cr
\itemize{
\item  YEAR
\item  STRATUM
\item  STRATUM_AREA
\item  other user-defined factors
\item  numStations
\item  numHauls
\item  numNonZeroHauls
\item  numIndivs
\item  totABUNDANCE = estimated abundance (by stratum)
\item  stdABUNDANCE = std deviation of estimated abundance (by stratum)
\item  cvABUNDANCE  = cv of estimated abundance (by stratum)
\item  totBIOMASS = estimated biomass (by stratum)
\item  stdBIOMASS = std deviation of estimated biomass (by stratum)
\item  cvBIOMASS  = cv of estimated biomass (by stratum)
}
}
\description{
This function calculates abundance, and biomass by stratum from cpue (by survey station or by haul).
}
\details{
If tbl_cpue is NULL, the user is prompted to enter a csv file with cpue info. \cr
\cr Other notes: \cr
\itemize{
  \item Area is in square nautical miles
  \item Abundance is in 10^6 indivs
  \item Biomass   is in 10^3 mt
}
\cr If \code{useStratumArea} is true, the stratum area is used to expand mean cpue to stratum abundance/biomass.
If it is false, the sum of STATION_AREAs associated with the stations included in the stratum is used for
the expansion (the sum of STATION_AREA across all standard stations in a stratum is close, but not exactly
equal to the STRATUM_AREA for that stratum).
}
