% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEffectiveN.SurveyData.NMFS.R
\name{calcEffectiveN.SurveyData.NMFS}
\alias{calcEffectiveN.SurveyData.NMFS}
\title{Calculate and plot effective N's for NMFS survey size compositions using a bootstrapping approach}
\usage{
calcEffectiveN.SurveyData.NMFS(
  loadRZCs = NULL,
  outDir = "./OutputFiles",
  title = "Effective N for NMFS Survey Size Compositions",
  bootN = 100,
  species = "BTC",
  minYr = 1975,
  maxYr = 2017,
  minSize = 25,
  maxSize = 185,
  binSize = 5,
  lglZ = 125,
  bySex = TRUE,
  byMaturity = TRUE,
  byShellCondition = FALSE,
  output_format = "pdf_document",
  clean = TRUE,
  verbosity = 0
)
}
\arguments{
\item{loadRZCs}{- NULL to calculate and plot resampled size compositions, or path to RData file to re-plot a saved set}
}
\description{
Function to calculate and plot effective N's for NMFS survey size compositions using a bootstrapping approach.
}
\details{
Uses rmd/calcEffectiveN.SurveyData.NMFS.Rmd to produce a report file. Saves results in outDir/RZCs.RData.
}
