% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResampledACDs.R
\name{calcResampledACDs}
\alias{calcResampledACDs}
\title{TODO: finish implementing this.}
\usage{
calcResampledACDs(
  tbl_strata,
  tbl_hauls = NULL,
  tbl_indivs = NULL,
  N = 100,
  byEW166 = FALSE,
  byEBS = TRUE,
  bySex = FALSE,
  byShellCondition = FALSE,
  byMaturity = FALSE,
  minSize = 25,
  maxSize = 185,
  truncate.low = TRUE,
  truncate.high = FALSE,
  verbosity = 0
)
}
\arguments{
\item{tbl_strata}{: data frame from call to \code{\link{selectStrata.TrawlSurvey}} [required]}

\item{tbl_hauls}{: dataframe from call to \code{\link{selectHauls.TrawlSurvey}} [required only if tbl_cpue not given]}

\item{tbl_indivs}{: dataframe from call to \code{\link{selectIndivs.TrawlSurvey}} (or crab survey filename, or NULL) [required only if tbl_cpue not given]}

\item{N}{: number of resamples/year}

\item{byEW166}{: flag (T/F) to aggregate size comps to EW166}

\item{byEBS}{: flag (T/F) to aggregate size comps to the EBS}

\item{bySex}{: flag (T/F) to calc by sex}

\item{byShellCondition}{: flag (T/F) to calc by shell condition}

\item{byMaturity}{: flag (T/F) to calc by maturity state}

\item{minSize}{: minimum size to include in calculations}

\item{maxSize}{: maximum size to include in calculations}

\item{truncate.low}{: flag (T/F) to exclude individuals smaller than minSize}

\item{truncate.high}{: flag (T/F) to exclude individuals larger than maxSize}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
data frame with ACDs by stratum. Columns are \cr
\itemize{
\item  i - resampling index (i=0 denotes the obsesrved size comps)
\item  YEAR
\item  STRATUM
\item  STRATUM_AREA
\item  other user-defined factors (e.g., sex, shell_condition)
\item  numStations
\item  numHauls
\item  numNonZeroHauls
\item  numIndivs
\item  totABUNDANCE = estimated abundance-by-size (by stratum)
\item  totBIOMASS = estimated biomass-by-size (by stratum)
}
}
\description{
Function to resample haul data to bootstrap aggregated catch data by year/stratum from AFSC trawl survey data
}
\details{
This function performs two-stage resampling (hauls w/in strata, individuals w/in hauls)
to calculate a set of resampled \strong{size compositions} (<-FIX THIS!) by year. \cr
\cr
Other notes: \cr
\itemize{
\item{Area is in square nautical miles}
\item{Abundance is in 10^6 indivs}
\item{Biomass   is in 10^3 mt}
}
}
