% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSizeComps.EW166.R
\name{calcSizeComps.EW166}
\alias{calcSizeComps.EW166}
\title{Calculate abundance, size comps size comps by year, east/west of 166W, and other factors from a size comps-by-stratum data frame or csv file.}
\usage{
calcSizeComps.EW166(tbl = NULL,
  strata_toEW166 = Codes.TrawlSurvey()[["strata.EW166"]], export = TRUE,
  out.csv = "SurveySizeComps.EW166.csv", out.dir = NULL, verbosity = 1)
}
\arguments{
\item{tbl}{: data frame with size comps by stratum info from call to \code{\link{calcSizeComps.ByStratum}} or csv file with size comps by stratum info, or NULL}

\item{strata_toEW166}{: data frame w/ conversion from original strata to EW166 strata}

\item{export}{: boolean flag to write results to csv file}

\item{out.csv}{: output file name}

\item{out.dir}{: output file directory}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
Dataframe w/ estimates of abundance, size comps size comps by year, strata as east/west of 166W, and other factors. Columns are \cr
\itemize{
\item  YEAR            = survey year
\item  STRATUM         = 'EAST' or 'WEST' of 166W
\item  STRATUM_AREA    = area of stratum
\item  other user-defined factors (e.g., sex, shell_condition)
\item  SIZE
\item  numStations     = number of stations included
\item  numHauls        = number of hauls included
\item  numNonZeroHauls = number of hauls included
\item  numIndivs       = number of individuals sampled
\item  totABUNDANCE    = estimated abundance
\item  totBIOMASS      = estimated size comps
}
}
\description{
This function estimates size comps for abundance, size comps by year, east/west of 166W from a size comps-by-stratum data frame or csv file.
}
\details{
Note: if tbl and in.csv are both NULL, the user is prompted to enter a csv file with size comps by stratum info. \cr
Notes: \cr
\itemize{
\item   Area is in square nautical miles
\item   Abundance is in 10^6 individuals
\item   Biomass   is in 10^3 mt
}
}
