% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSSs.R
\name{calcSSs}
\alias{calcSSs}
\title{Calculate sample sizes for survey data}
\usage{
calcSSs(dfrACD, colName = "numIndivs", totVal = 200)
}
\arguments{
\item{dfrACD}{: dataframe with aggregated catch data}

\item{colName}{: column name with sample sizes}

\item{totVal}{: total SS for relative SS calculations}
}
\value{
list with elements \code{totSS} and \code{relSS}. Both are dataframes.
}
\description{
Function to calculate total and relative sample sizes
}
\details{
None.
}
