\name{Codes.TrawlSurvey}
\alias{Codes.TrawlSurvey}
\title{Provides a list of dataframes suitable forconverting from survey codes to
assessment codes and descriptions.}
\usage{
Codes.TrawlSurvey()
}
\value{
a list with the following elements: sex : table to convert
between survey and assessment sex codes shell_condition :
table to convert between survey and assessment shell
condition codes clutch_size : table to convert between
survey and assessment clutch size codes strata.BTC :
strata.orig.BTC : strata.EW166 :
}
\description{
This function provides a list of dataframes suitable for
converting from survey codes to assessment codes and
descriptions
}
\details{

}

