% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFormat.BSFRF2NMFS.R
\name{convertFormat.BSFRF2NMFS}
\alias{convertFormat.BSFRF2NMFS}
\title{Convert BSFRF data to (almost) NMFS CRABHAUL_DATA format}
\usage{
convertFormat.BSFRF2NMFS(tbl, types = c("SBS", "IDX"),
  verbosity = TRUE)
}
\arguments{
\item{tbl}{- filename for BSFRF  csv file to read or dataframe from reading csv file}

\item{types}{- BSFRF study types to select ("SBS", "IDX", or both)}

\item{verbose}{- flag to print debug info}
}
\value{
dataframe "almost" in CRABHAUL_DATA format.
}
\description{
Function to convert BSFRF data to (almost) NMFS CRABHAUL_DATA format.
}
\details{
returned dataframe has columns
\itemize{
  \item {AKFIN_SURVEY_YEAR}
  \item {HAULJOIN}
  \item {VESSEL}
  \item {CRUISE}
  \item {HAUL}
  \item {HAUL_TYPE}
  \item {START_DATE}
  \item {AKFIN_SURVEY_DATE}
  \item {START_HOUR}
  \item {MID_LATITUDE}
  \item {MID_LONGITUDE}
  \item {GIS_STATION}
  \item {BOTTOM_DEPTH}
  \item {GEAR_TEMPERATURE}
  \item {AREA_SWEPT_VARIABLE}
  \item {SPECIES_CODE}
  \item {SPECIES_NAME}
  \item {SEX}
  \item {WIDTH}
  \item {SHELL_CONDITION}
  \item {EGG_COLOR}
  \item {EGG_CONDITION}
  \item {CLUTCH_SIZE}
  \item {CHELA_HEIGHT}
  \item {CALCULATED_WEIGHT}
  \item {WEIGHT}
  \item {SAMPLING_FACTOR}
  \item {CPUE_NUM}
}
}
