% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataTables.R
\name{createDataTables}
\alias{createDataTables}
\title{Create strata, haul and individual data tables (i.e., dataframes) from strata and crabhaul files}
\usage{
createDataTables(species = c("Tanner crab", "Snow crab", "BKC", "RKC"),
  surveyType = c("NMFS", "BSFRF"), bsfrfTypes = c("SBS", "IDX"),
  strataTbl = NULL, strataFile = "", strataType = "2015",
  haulFiles = "", HaulTypes = NULL, minYr = 2016, maxYr = 2017,
  sex = "ALL", maturity = "ALL", shell_condition = "ALL",
  minSize = 0, maxSize = Inf, calcMaleMaturity = FALSE,
  verbosity = 0)
}
\arguments{
\item{species}{: character string indicating species (i.e., "Tanner crab","Snow crab","BKC", or "RKC")}

\item{surveyType}{: survey type ("NMFS" or "BSFRF")}

\item{bsfrfTypes}{: vector of BSFRF types to include ("SBS" and/or "IDX")}

\item{strataTbl}{: dataframe previously created by call to \code{\link{selectStrata.TrawlSurvey}}}

\item{strataFile}{: path to strata file (req'd if strataTbl is NULL)}

\item{strataType}{: "2015"}

\item{haulFiles}{: vector of paths to crabhaul files}

\item{HaulTypes}{: haul types to include (NULL selects all)}

\item{minYr}{: min year to include}

\item{maxYr}{: max year to include}

\item{sex}{: sexes to include ("MALE", "FEMALE", or "ALL")}

\item{maturity}{: maturity states to include ("IMMATURE", "MATURE", "UNDETERMINED" , or "ALL")}

\item{shell_condition}{: shell conditions to include ("NEW", "OLD", "ALL")}

\item{minSize}{: min size (CW) to include}

\item{maxSize}{: max size (CW) to include}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
a list with the following elements: \cr
\itemize{
\item {dfrStrataData} {- dataframe with stratum information; see \code{\link{selectStrata.TrawlSurvey}}}
\item {dfrHaulData}   {- dataframe with haul information; see \code{\link{selectHauls.TrawlSurvey}}}
\item {dfrIndivData}  {- dataframe with individual crab data; see \code{\link{selectIndivs.TrawlSurvey}}}
}

A list with three dataframes. See Details.
}
\description{
Function to create strata, haul and individual data tables (i.e., dataframes) from strata and crabhaul files.
}
\details{

}
