% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputFile.SurveyData.NMFS.R
\name{createInputFile.SurveyData.NMFS}
\alias{createInputFile.SurveyData.NMFS}
\title{Creates a survey input file for TCSAM02}
\usage{
createInputFile.SurveyData.NMFS(fn, cutpts, dfrACD, dfrSS, dfrZCs,
  verbose = FALSE)
}
\arguments{
\item{fn}{- output file name}

\item{cutpts}{- vector of cutpoints for the input size compositions}

\item{dfrACD}{- dataframe of aggregated catch data}

\item{dfrSS}{- dataframe of input sample sizes corresponding to the size comps}

\item{dfrZCs}{- dataframe of size comps}

\item{verbose}{- flag to print debugging info}
}
\description{
This function that creates a survey input file for TCSAM02.
}
\details{
This function that creates a survey input file for TCSAM02.
}
