% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCalcs_CPUEbyStation.R
\name{doCalcs_CPUEbyStation}
\alias{doCalcs_CPUEbyStation}
\title{Calculate cpue by survey station and other factors from strata and haul files}
\usage{
doCalcs_CPUEbyStation(species = c("Tanner crab", "Snow crab", "BKC",
  "RKC"), surveyType = c("NMFS", "BSFRF"), bsfrfTypes = c("SBS",
  "IDX"), strataTbl = NULL, strataFile = "", strataType = "2015",
  haulFiles = "", HaulTypes = NULL, minYr = 2016, maxYr = 2017,
  sex = "ALL", maturity = "ALL", shell_condition = "ALL",
  minSize = 0, maxSize = Inf, calcMaleMaturity = FALSE,
  cutpts = seq(from = 25, to = 185, by = 5), truncate.low = TRUE,
  truncate.high = FALSE, aggBySex = FALSE, aggByMaturity = FALSE,
  aggByShellCondition = FALSE, aggBySize = FALSE, verbosity = 0)
}
\arguments{
\item{species}{: character string indicating species (i.e., "Tanner crab","Snow crab","BKC", or "RKC")}

\item{surveyType}{: survey type ("NMFS" or "BSFRF")}

\item{bsfrfTypes}{: vector of BSFRF types to include ("SBS" and/or "IDX")}

\item{strataTbl}{: dataframe previously created by call to \code{\link{selectStrata.TrawlSurvey}}}

\item{strataFile}{: path to strata file (req'd if strataTbl is NULL)}

\item{strataType}{: "2015"}

\item{haulFiles}{: vector of paths to haul files}

\item{HaulTypes}{: haul types (NULL selects all)}

\item{minYr}{: min year to include}

\item{maxYr}{: max year to include}

\item{sex}{: sexes to include ("MALE", "FEMALE", or "ALL")}

\item{maturity}{: maturity states to include ("IMMATURE", "MATURE", "UNDETERMINED" , or "ALL")}

\item{shell_condition}{: shell conditions to include ("NEW", "OLD", "ALL")}

\item{minSize}{: min size (CW) to include}

\item{maxSize}{: max size (CW) to include}

\item{calcMaleMaturity}{: flag (T/F) to calculate male maturity using the Rugolo/Turnock maturity ogive}

\item{cutpts}{: vector of cutpoints to create size bins from}

\item{truncate.low}{: flag (T/F) to exclude individuals smaller than minSize}

\item{truncate.high}{: flag (T/F) to exclude individuals larger than maxSize}

\item{aggBySex}{: flag to agregate CPUE over sexes}

\item{aggByMaturity}{: flag to aggregate of maturity states}

\item{aggByShellCondition}{: flag to aggregate over shell conditions}

\item{aggBySize}{: flag to aggregate over sizes}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
A dataframe of cpue (numbers and weight) by haul. See Details.
}
\description{
Function to calculate cpue by survey sstation and other factors (e.g., sex) from strata and haul files.
}
\details{
The returned dataframe has the following columns: \cr
\itemize{
\item   YEAR
\item   STRATUM
\item   GIS_STATION
\item   HAULJOIN
\item   LONGITUDE
\item   LATITUDE
\item   SEX
\item   MATURITY
\item   SHELL_CONDITION
\item   SIZE
\item   numHauls
\item   numNonZeroHauls
\item   numIndivs
\item   numCPUE
\item   wgtCPUE
}
\cr Other notes: \cr
\itemize{
\item   CPUE in numbers is in no/(sq. nm.)
\item   CPUE in weight  is in mt/(sq. nm.)
} \cr
}
