% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCalcs_SSs.R
\name{doCalcs_SSs}
\alias{doCalcs_SSs}
\title{Function to calculate sample sizes at various levels of spatial aggregation from station, haul and individual crab info.}
\usage{
doCalcs_SSs(tbl_strata, tbl_hauls, tbl_indivs,
  averageHaulsByStation = TRUE, aggBySex = FALSE,
  aggByMaturity = FALSE, aggByShellCondition = FALSE,
  dropLevels = list(SEX = c("MISSING", "HERMAPHRODITIC")),
  colName = "numIndivs", totVal = 200, verbosity = verbosity)
}
\arguments{
\item{tbl_strata}{- dataframe with survey strata info (output from \code{\link{selectStrata.TrawlSurvey}})}

\item{tbl_hauls}{- dataframe with hauls info (output from \code{\link{selectHauls.TrawlSurvey}})}

\item{tbl_indivs}{- dataframe with individual crab info (output from \code{\link{selectIndivs.TrawlSurvey}})}

\item{averageHaulsByStation}{- TRUE (average hauls by station prior to calculating stratum-level quantities)}

\item{aggBySex}{- flag to agregate CPUE over sexes}

\item{aggByMaturity}{- flag to aggregate of maturity states}

\item{aggByShellCondition}{- flag to aggregate over shell conditions}

\item{dropLevels}{- NULL, or list (by factor name) of vectors of factor values to drop (see help for \code{wtsUtilities::dropLevels})}

\item{colName}{- column name with sample sizes}

\item{totVal}{- total SS for relative SS calculations}

\item{verbosity}{- integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
a list with elements \code{ss.EBS} and \code{relSS.EBS}. \code{ss.EBS} is a dataframe with columns: \cr
\itemize{
\item {year}     {- survey year}
\item {ssTot}    {- total sample size}
} \cr
\code{relSS.EBS} is a dataframe with the following columns: \cr
\itemize{
\item {YEAR}             {- }
\item {SEX}              {- }
\item {MATURITY}         {- }
\item {SHELL_CONDITION}  {- }
\item {relSS}            {relative sample size}
}
}
\description{
Function to calculate sample sizes from station, haul and individual crab info.
}
\details{
None.
}
