% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exportAggregatedCatchData.R
\name{exportAggregatedCatchData}
\alias{exportAggregatedCatchData}
\title{Export aggregated catch data to stock assessment format.}
\usage{
exportAggregatedCatchData(acd, facs = "", dropLevels = NULL,
  var = c("ABUNDANCE", "BIOMASS"), export = FALSE,
  out.csv = paste("SurveyACD", var[1], "csv", sep = "."), out.dir = NULL,
  verbosity = 1)
}
\arguments{
\item{acd}{- dataframe from call to one of the calcBiomass... functions}

\item{facs}{- factors to include in output}

\item{dropLevels}{- factor levels to drop}

\item{var}{- variable type to export (abundance or biomass)}

\item{export}{- boolean flag to write results to csv file}

\item{out.csv}{- output file name}

\item{out.dir}{- output file directory}

\item{verbosity}{- integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
dataframe in stock assessment format
}
\description{
Function to export aggregated catch data to stock assessment format.
}
\details{


Factors (e.g., 'SEX', 'MATURITY', 'SHELL_CONDITION') in the input dataframe that are
missing from the 'facs' vector will be aggregated over by summing. STRATUM is \bold{not} regarded as
a potential factor. The time series for each factor level combination is exported.
Distinct levels of each factor can be dropped from the
exported dataframe by seting dropLevels to a list with names corresponding to
factor columns and values being vectors of factor levels to drop.

Output column order will be STRATUM, facs, YEAR, value, cv
}

