% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSizeComps.fromCPUEbyStation.R
\name{extractSizeComps.fromCPUEbyStation}
\alias{extractSizeComps.fromCPUEbyStation}
\title{Extract size compositions from size-specific CPUE data by station}
\usage{
extractSizeComps.fromCPUEbyStation(tbl_cpue = NULL, cutpts = seq(from = 25,
  to = 185, by = 5), export = FALSE,
  out.csv = "SurveySizeComps.CPUE.ByStratum.csv", out.dir = NULL,
  verbosity = 0)
}
\arguments{
\item{tbl_cpue}{: data frame from call to \code{\link{calcCPUE.ByStation}} or filename for csv file}

\item{cutpts}{: vector of cutpoints to create size bins from}

\item{export}{: boolean flag to write results to csv file}

\item{out.csv}{: output file name}

\item{out.dir}{: output file directory}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
data frame with size comps by stratum. Columns are \cr
\itemize{
\item  YEAR
\item  STRATUM
\item  STRATUM_AREA
\item  other user-defined factors (e.g., sex, shell_condition)
\item  SIZE
\item  numStations
\item  numHauls
\item  numNonZeroHauls
\item  numIndivs
\item  avgNumCPUE = average cpue-by-size (numbers/sq. nautical mile) by stratum
\item  stdNumCPUE = standard deviation in cpue-by-size (numbers/sq. nautical mile) by stratum
\item  cvNumCPUE = cv in  cpue-by-size (numbers/sq. nautical mile) by stratum
\item  avgWgtCPUE = estimated cpue-by-size (mt/sq. nautical mile) by stratum
\item  stdWgtCPUE = standard deviation in cpue-by-size (mt/sq. nautical mile) by stratum
\item  cvWgtCPUE = cv in  cpue-by-size (mt/sq. nautical mile) by stratum
}
}
\description{
This function extracts size compositions from size-specific CPUE data by NMFS station.
}
\details{
If tbl_cpue is not a dataframe, it can be a cpue csv file or NULL.\cr
\cr
Other notes: \cr
\itemize{
  \item{Area is in square nautical miles}
  \item{avgNumCPUE is in indivs/square nautical mile}
  \item{avgBioCPUE is in mt/square nautical mile}
}
}
