% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gClip.R
\name{gClip}
\alias{gClip}
\title{Clip a set of polygons using a bounding box}
\usage{
gClip(shp, bb)
}
\arguments{
\item{shp}{- set of polygons (from call to readOGR)}

\item{bb}{- bounding box (see details)}
}
\value{
- spatial dataframe of clipped polygons
}
\description{
Function to clip a set of polygons using a bounding box
}
\details{


The bounding box can be one of the following (see extent{raster}):\cr
1. vector of coordinates given as xmin, xmax, ymin, ymax\cr
2. vector of coordinates given as (with names): "left","bottom","top","right"\cr
3. matrix with columns corresponding to "x","y" and rows to "lower left", "upper right"\cr
4. dataframe with columns corresponding to "x","y" and rows to "lower left", "upper right"\cr
\cr
Also, the bounding box must be in the same coordinate system as th set of polygons.
}

