% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stamenmap.R
\name{get_stamenmap}
\alias{get_stamenmap}
\title{Retrieve a stamenmap (replaces ggmap::get_stamenmap)}
\usage{
get_stamenmap(bbox = c(left = -180, bottom = 54, right = -155, top = 63),
  zoom = 6, maptype = c("terrain", "terrain-background", "terrain-labels",
  "terrain-lines", "toner", "toner-2010", "toner-2011", "toner-background",
  "toner-hybrid", "toner-labels", "toner-lines", "toner-lite", "watercolor"),
  land = "black", water = "white", crop = TRUE, messaging = FALSE,
  urlonly = FALSE, filename = "ggmapTemp", color = c("color", "bw"), ...)
}
\arguments{
\item{bbox}{- bounding box coordinates (left, bottom, right, top) (default is for EBS)}

\item{zoom}{- zoom level fordownloading tiles (1- 18)}

\item{maptype}{- "terrain", "watercolor" or "toner"}

\item{land}{- color for land when maptype is "toner"}

\item{water}{- color for water when maptype is "toner"}

\item{crop}{- crop tiles to bbox}

\item{messaging}{- report messages}

\item{urlonly}{- return only the url}

\item{filename}{- temporary file name}

\item{color}{- 'color' or 'bw'}

\item{...}{- other inputs to ggmap::get_stamenmap}
}
\value{
ggmap object

object to use with ggmap
}
\description{
Function to retrieve a stamenmap (replaces ggmap::get_stamenmap)
}
\details{
Requires internet connection.
}
