% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMapData.R
\name{getMapData}
\alias{getMapData}
\title{Get map data objects associated with this package}
\usage{
getMapData(type = c("ggmap", "shape"), name = c("EBS.stamen.toner.z06",
  "EBS.stamen.toner.z07", "EBS.polys.land", "EBS.polys.surveyblocks",
  "EBS.lines.bathy"), verbosity = 0)
}
\arguments{
\item{type}{- object type. one of ["ggmap","shape"]}

\item{name}{- object name. see details.}
}
\value{
ggmap or dataframe object from package data
}
\description{
Function to return map data objects associated with this package.
}
\details{
'name' must be one of
\itemize{
 \item "EBS.stamen.toner.z06"
 \item "EBS.stamen.toner.z07"
 \item "EBS.polys.land"
 \item "EBS.polys.surveyblocks"
 \item "EBS.lines.bathy"
}
}
