% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaturityOgives.Females.R
\name{getMaturityOgives.Females}
\alias{getMaturityOgives.Females}
\title{Function to extract female maturity ogives from a size composition dataframe}
\usage{
getMaturityOgives.Females(
  dfrZCs = NULL,
  showPlot = TRUE,
  export = FALSE,
  out.csv = "MaturityOgives.Females.csv",
  out.dir = NULL,
  verbosity = 0,
  ...
)
}
\arguments{
\item{dfrZCs}{: dataframe with new shell female size compositions by maturity state}

\item{showPlot}{- flag to show plot of ogives}

\item{export}{- boolean flag to export results to csv file}

\item{out.csv}{- name of output csv file                    (ignored if NULL)}

\item{out.dir}{- base path for output csv file              (set to folder of input csv file or current working directory)}

\item{verbosity}{- flags for intermediate output}

\item{...}{- parameters passed to plotMaturityOgives(...)}
}
\value{
dataframe with columns: \cr
\itemize{
\item  YEAR
\item  STRATUM
\item  SEX
\item  SIZE
\item  numIndivs
\item  ogive
}
}
\description{
Function to extract female maturity ogives from a size composition dataframe.
}
\details{
the input size composition dataframe should be in the output format of the calcSizeComps... functions.\cr
}
