% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_GetBasemapLayers.R
\name{gg_GetBasemapLayers}
\alias{gg_GetBasemapLayers}
\title{Create basemap layers for maps based on the \pkg{ggplot2} package}
\usage{
gg_GetBasemapLayers(
  sf_land = NULL,
  sf_bathym = NULL,
  sfs_survey = NULL,
  final_crs = wtsGIS::get_crs("AlaskaAlbers"),
  bbox = wtsGIS::getStandardBBox("EBS"),
  bw = FALSE,
  colors = list(land = "green", bathym = "blue", grid = "dark grey", stations = "red"),
  size = 0.01
)
}
\arguments{
\item{sf_land}{- a spatial layer representing land (or NULL)}

\item{sf_bathym}{- a spatial layer representing bathymetry (or NULL)}

\item{sfs_survey}{-  a list of spatial layers representing the survey grid and survey stations (or NULL)}

\item{final_crs}{- string representation of CRS (default = WGS84) used for ALL shapefiles}

\item{bbox}{- a \code{sf::bbox} bounding box}

\item{bw}{- flag (T/F) to create layers in black & white (TRUE) or color (FALSE)}

\item{colors}{- list of colors to use for "land","bathym", "grid" and "stations"}

\item{size}{- point size for the station locations}
}
\value{
- a list of \pkg{ggplot2} objects
}
\description{
This function creates basemap layers for maps based on the \pkg{ggplot2} package
}
\details{
Uses functions from package [wtsGIS].

The returned list has elements:
\itemize{
  \item{bathym}
  \item{land}
  \item{grid}
  \item{stations}
  \item{map_scale}
  \item{theme}
}
}
