% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.ggmaps.R
\docType{data}
\name{ggmap.07.EBS}
\alias{ggmap.07.EBS}
\title{A Stamen toner map for the EBS at zoom = 7}
\format{A ggmap raster object.}
\usage{
ggmap.07.EBS
}
\description{
A ggmap raster object reflecting a Stamen background toner map of EBS 
at zoom = 7. The bounding box is left=-180, bottom=54, right=-155, top=63.
}
\keyword{datasets}

