% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.data.R
\docType{data}
\name{ggmap.EBS.stamen.toner.z06}
\alias{ggmap.EBS.stamen.toner.z06}
\title{A Stamen toner map for the EBS at zoom = 6}
\format{A ggmap raster object.}
\usage{
ggmap.EBS.stamen.toner.z06
}
\description{
A ggmap raster object reflecting a Stamen background toner map of EBS 
at zoom = 6. The bounding box is left=-180, bottom=54, right=-155, top=63.
}
\keyword{datasets}

