% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisCreateBaseMap.R
\name{gisCreateBaseMap}
\alias{gisCreateBaseMap}
\title{Create a basemap layer for maps based on the tmap package}
\usage{
gisCreateBaseMap(
  layer.land = NULL,
  layer.bathym = NULL,
  layers.survey = NULL,
  gisPath = NULL,
  shapeFile.bathymetry = NULL,
  shapeFile.land = NULL,
  shapeFiles.survey = list(grid = NULL, stations = NULL),
  strCRS = wtsGIS::getCRS("WGS84"),
  boundingbox = list(bottomleft = list(lon = -179, lat = 54), topright = list(lon =
    -157, lat = 62.5)),
  colors.bathym = "darkblue",
  points.size = 0.01,
  as.sf = TRUE
)
}
\arguments{
\item{layer.land}{- a tmap layer representing land (or NULL)}

\item{layer.bathym}{- a tmap layer representing bathymetry (or NULL)}

\item{layers.survey}{-  a list of tmap layers representing the survey grid and survey stations (or NULL)}

\item{gisPath}{- path to top level folder for shapefiles}

\item{shapeFile.bathymetry}{- bathymetry shapefile name}

\item{shapeFile.land}{- land shapefile name}

\item{shapeFiles.survey}{- list (grid, stations) of shapfile names to define survey stations}

\item{strCRS}{- string representation of CRS (default = WGS84) used for ALL shapefiles}

\item{boundingbox}{- a bounding box (list with elements "bottomleft" and "topright")}

\item{colors.bathym}{- color for the bathymetry}

\item{points.size}{- size for the station locations}

\item{as.sf}{- flag to create spatial datasets as sf (simple features) objects (TRUE) or as sp objects (FALSE)}
}
\value{
- basemap layer in WGS84 based on the tmap package
}
\description{
This function creates a basemap layer for maps based on the tmap package
}
\details{
Uses \code{wtsGIS::tmap.CreateLayerFromShapefile}.
}
