% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisCreateSurveyMap.CPUE.R
\name{gisCreateSurveyMap.CPUE}
\alias{gisCreateSurveyMap.CPUE}
\title{Create a tmap-style map of survey cpue as bubbles}
\usage{
gisCreateSurveyMap.CPUE(basemap, cpue, raster = NULL, cpueCol = "CPUE",
  cpueColor = "blue", cpueAlpha = 0.6, maxCPUE = NULL,
  cpueLabel = "CPUE", cpueLegend = TRUE, rasterCol = "z",
  rasterPalette = NULL, rasterAlpha = 0.8, rasterLabel = "",
  rasterBreaks = NULL, rasterLegend = TRUE,
  panelLabel = "Survey CPUE")
}
\arguments{
\item{basemap}{- tmap-style basemap}

\item{cpue}{- spatial dataframe with cpue data}

\item{raster}{- raster object with environmental data}

\item{cpueCol}{- name of column in cpue dataframe to plot as bubbles}

\item{cpueColor}{- color for cpue bubbles}

\item{cpueAlpha}{- alpha (transparency) for cpue bubbles}

\item{maxCPUE}{- max value for cpue scale}

\item{cpueLabel}{- label for cpue scale}

\item{cpueLegend}{- TRUE/FALSE to show cpue scale}

\item{rasterCol}{- name of raster data column}

\item{rasterPalette}{- palette for raster}

\item{rasterAlpha}{- alpha (transparency) for raster}

\item{rasterLabel}{- label for raster scale}

\item{rasterBreaks}{- breaks for raster scale}

\item{rasterLegend}{- TRUE/FALSE to show raster scale}

\item{panelLabel}{- label for map panel}
}
\value{
tmap map object
}
\description{
Function to create a tmap-style map of survey cpue as bubbles.
}
\details{
Uses the tmap package.
}
