% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisGetSurveyGridLayers.R
\name{gisGetSurveyGridLayers}
\alias{gisGetSurveyGridLayers}
\title{Retrieve the EBS bottom trawl survey grid as pre-packaged polygon and point \pkg{sf} dataframes}
\usage{
gisGetSurveyGridLayers()
}
\value{
a 2-element list with the EBS trawl survey grid and station information as \pkg{sf} dataframes.
}
\description{
This function retrieves pre-packaged \pkg{sf} polygon and point dataframes for the
EBS bottom trawl survey.
}
\details{
None.
}
