% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotACD.R
\name{plotACD}
\alias{plotACD}
\title{Plot time series of aggregated catch data}
\usage{
plotACD(acd, type = c("abundance", "biomass"), factors = NULL,
  faceting = NULL, facetsScale = "fixed", facetsDrop = FALSE,
  xlab = "year", ylab = NULL, ci = 0.8, ci.type = "normal",
  xlims = NULL, ylims = NULL, ggtheme = theme_grey(),
  jitter = TRUE, showPlots = TRUE, verbosity = 0)
}
\arguments{
\item{acd}{- dataframe with aggregated catch data (e.g., abundance or biomass)}

\item{type}{- "abundance" or "biomass"}

\item{factors}{- column names to use as factors for plots}

\item{faceting}{- faceting formula (depnds on ggplot2 version)}

\item{facetsScale}{- value for \code{scale} parameter passed to \code{facet_grid()} ("fixed","free_x","free_y","free")}

\item{facetsDrop}{- value for \code{drop} parameter passed to \code{facet_grid()} (drop missing facet levels)}

\item{xlab}{- x-axis label ("year")}

\item{ylab}{- y-axis label}

\item{ci}{- confidence interval for error bars (e.g., 0.95)}

\item{ci.type}{- confidence interval type ('normal', 'lognormal')}

\item{xlims}{- x-axis limits (calculated internally if NULL)}

\item{ylims}{- y-axis limits (calculated internally if NULL)}

\item{ggtheme}{- ggplot2 theme}

\item{jitter}{- flag to jitter points}

\item{showPlots}{- flag to show plots immediately}

\item{verbosity}{- flag (>0) to print diagnostic output}
}
\value{
list of ggplot2 plot objects
}
\description{
Function to plot time series of aggregated catch data.
}
\details{
The time series for each factor level combination is plotted
separately. \code{aggregateCatchData()} should be used to be drop
undesired factor levels prior to plotting.

One plot is created for each distinct level of 'STRATUM'.
}
