% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAggregatedCatchData.R
\name{plotAggregatedCatchData.ByStratum}
\alias{plotAggregatedCatchData.ByStratum}
\title{Plot time series of aggregated catch data.}
\usage{
plotAggregatedCatchData.ByStratum(acd, facs = "", dropLevels = NULL,
  var = c("ABUNDANCE", "BIOMASS"), ci = 0.8, ci.type = "normal",
  rng = NULL, ggtheme = theme_grey(), ncol = 2, nrow = 5,
  showPlots = TRUE)
}
\arguments{
\item{acd}{- dataframe from call to one of the calcBiomass... functions}

\item{facs}{- factors to plot by}

\item{dropLevels}{- factor levels to drop from plots}

\item{var}{- variable type to plot (abundance or biomass)}

\item{ci}{- confidence interval for error bars (e.g., 0.95)}

\item{ci.type}{- confidence interval type ('normal', 'lognormal')}

\item{rng}{- y-axis range (calculated internally if NULL)}

\item{ggtheme}{- ggplot2 theme}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlots}{- flag to show plots immediately}
}
\value{
list of ggplot2 plot objects
}
\description{
Function to plot time series of aggregated catch data.
}
\details{
The time series for each factor level combination is plotted
separately. Distinct levels of each factor can be dropped from the
final plot by seting dropLevels to a list with names corresponding to 
factor columns and values being vectors of factor levels to drop.

One plot is created for each distinct level of 'STRATUM'.
}

