% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAZC.R
\name{plotAZC}
\alias{plotAZC}
\title{Plot aggregated size comps from survey data as bars}
\usage{
plotAZC(
  zcs,
  byStratum = TRUE,
  multipliers = NULL,
  zlab = "Abundance (millions)",
  zlims = NULL,
  xlims = NULL,
  ylims = NULL,
  ggtheme = theme_grey(),
  ncol = 2,
  nrow = 5,
  showPlots = TRUE,
  verbosity = 0
)
}
\arguments{
\item{zcs}{\itemize{
\item dataframe from call to aggregateSizeComps(...)
}}

\item{byStratum}{\itemize{
\item flag to plot size comps separately by stratum
}}

\item{multipliers}{\itemize{
\item list by factor of multipliers for factor levels
}}

\item{zlab}{\itemize{
\item 'z'-axis label
}}

\item{zlims}{\itemize{
\item z-axis range (calculated internally if NULL)
}}

\item{xlims}{\itemize{
\item size limits (calculated internally if NULL)
}}

\item{ylims}{\itemize{
\item year limits (calculated internally if NULL)
}}

\item{ggtheme}{\itemize{
\item ggplot2 theme
}}

\item{ncol}{\itemize{
\item number of columns of plots per page
}}

\item{nrow}{\itemize{
\item number of rows of plots per page
}}

\item{showPlots}{\itemize{
\item flag to show plots immediately
}}

\item{verbosity}{\itemize{
\item integer flag (>0) to print diagnostic output
}}
}
\value{
list of ggplot2 plot objects
}
\description{
Function to plot aggregated size comps from survey data as bars.
}
\details{
'multipliers' is a list of lists, with each sublist has two elements, 'factors' and 'value'.
'factors' should be a list of factor=level pairs. 'value' should be the value to multiply by. Requires
\link[scales:oob]{scales::squish}.
}
