% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotChelaHeights.R
\name{plotChelaHeights}
\alias{plotChelaHeights}
\title{Plot chela height data}
\usage{
plotChelaHeights(dfrp, ratio = 0.18, aggYears = 5, rng = NULL, ncol = 2,
  showPlot = FALSE)
}
\arguments{
\item{dfrp}{- dataframe of chela height data from call to getChelaHeights(...)}

\item{ratio}{- default ratio for mature/immature determination}

\item{aggYears}{- number of years over which to aggregate}

\item{rng}{- y-axis range}

\item{ncol}{- number of coulmns to use for faceting}

\item{showPlot}{- flag to show plots immediately}
}
\value{
ggplot2 plot object to print
}
\description{
Function to plot chela height data
}
\details{

}

