% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGGMaps.Scalar.R
\name{plotGGMaps.Scalar}
\alias{plotGGMaps.Scalar}
\title{Plot maps of survey temperature (or other scalar) by year using ggmap and ggplot2}
\usage{
plotGGMaps.Scalar(dfr, scalar = "TEMP", min = NULL, max = NULL,
  truncateMin = TRUE, truncateMax = TRUE, map = getMapData("ggmap",
  "EBS.stamen.toner.z06"), bbox = c(left = -180, bottom = 54, right = -155,
  top = 63), mapcolors = list(land = "darkgreen", water = "lightblue"),
  bathymetry = getMapData("shape", "EBS.lines.bathy"),
  land = getMapData("shape", "EBS.polys.land"), layers = NULL,
  label = "Temperature (deg C)", ggtheme = theme_grey(), ncol = 1,
  nrow = 4, showPlot = FALSE, verbosity = 0)
}
\arguments{
\item{dfr}{- dataframe with columns YEAR, LATITUDE, LONGITUDE, VALUE}

\item{scalar}{- name of scalar column in dfr to plot}

\item{min}{- min value to plot (or NULL)}

\item{max}{- max value to plot (or NULL)}

\item{truncateMin}{- flag to remove values below min}

\item{truncateMax}{- flag to remove values above max}

\item{map}{- ggmap raster object for base map background}

\item{bbox}{- coordinates of bounding box for map (left, bottom, right, top)}

\item{mapcolors}{- list of colors (land, water) for base map background}

\item{bathymetry}{- fortified spatial dataframe or list(dsn=,layer=) specifying bathymetry shapefile to include on base map}

\item{land}{- fortified spatial dataframe or list(dsn=,layer=) specifying land shapefile to include on base map}

\item{layers}{- list of ggplot layers to include in base map}

\item{label}{- page title}

\item{ggtheme}{- ggplot2 theme}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlot}{- flag to show plots immediately}

\item{verbosity}{- integer flag (>0) to print intermediate info}

\item{returnLayerObjects}{- flag to return ggplot2 layer objects (instead of ggplot2 plot objects)}
}
\value{
list of ggplot2 plot or layer objects to print
}
\description{
Function to plot maps of survey temperature (or other scalar) by year using ggmap and ggplot2
}
\details{
none.
}
