% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaturityOgives.R
\name{plotMaturityOgives}
\alias{plotMaturityOgives}
\title{Plot maturity ogive data}
\usage{
plotMaturityOgives(
  dfrp,
  facet_grid = "dcd~.",
  ncol = 4,
  title = NULL,
  showPlot = FALSE
)
}
\arguments{
\item{dfrp}{- dataframe of maturity ogive data}

\item{facet_grid}{- formula to use for a faceted grid plot (default=NULL)}

\item{ncol}{- number of plot columns to use when facet_grid is NULL}

\item{title}{- title for plot}

\item{showPlot}{- flag to show plots immediately}
}
\value{
ggplot2 plot object to print
}
\description{
Function to plot maturity ogive data.
}
\details{
If "ratio" is NULL (not the default), then maturity status (immature/mature) will not be determined
(even if "discrim" is given) and the maturity state-specific linear model fits (CH ~ CW or ln{CH}~ln{CW})
will not be shown. If "ratio" is not NULL (the default), then the discriminant function (if "discrim" is not NULL)
or the given ratio will be used to charaterize maturity state on the basis of CH vs. CW.

If "facet_grid" is NULL, then plots are produced by facet_wrapping by YEAR. Otherwise, the faceting formula given by
"facet_grid" is used to create a faceted grid plot.
}
