% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResampledSizeComps.R
\name{plotResampledSizeComps}
\alias{plotResampledSizeComps}
\title{Plot resampled size compositions}
\usage{
plotResampledSizeComps(dfr, N = 200, scales = "free_y")
}
\arguments{
\item{dfr}{- dataframe with resampled size compositions}

\item{N}{- assumed effective sample size (for plotting)}

\item{scales}{- 'scales' parameter passed to facet_grid ("fixed" or "free_y")}
}
\description{
Function to plot resampled size compositions.
}
\details{
dfr should be an output from \code{\link{calcResampledSizeComps}}.
}
