% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSizeComps.AsLines.R
\name{plotSizeComps.asLines}
\alias{plotSizeComps.asLines}
\title{Plot time series of size comps as lines (not bars).}
\usage{
plotSizeComps.asLines(zcs, facs = "", dropLevels = NULL,
  var = c("totABUNDANCE", "totBIOMASS"), rng = NULL,
  ggtheme = theme_grey(), ncol = 2, nrow = 5, showPlots = TRUE)
}
\arguments{
\item{zcs}{- dataframe from call to one of the calcSizeComps... functions}

\item{facs}{- factors to plot by}

\item{dropLevels}{- factor levels to drop from plots}

\item{var}{- variable type to plot (abundance or biomass)}

\item{rng}{- y-axis range (calculated internally if NULL)}

\item{ggtheme}{- ggplot2 theme}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlots}{- flag to show plots immediately}
}
\value{
list of ggplot2 plot objects
}
\description{
Function to plot time series of size comps as lines (not bars).
}
\details{
The time series for each factor level combination is plotted
separately. Distinct levels of each factor can be dropped from the
final plot by seting dropLevels to a list with names corresponding to 
factor columns and values being vectors of factor levels to drop.

One plot is created for each distinct level of 'STRATUM'.
}

