% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSizeComps.ByStratum.R
\name{plotSizeComps.ByStratum}
\alias{plotSizeComps.ByStratum}
\title{Plot size comps from survey data as bars.}
\usage{
plotSizeComps.ByStratum(zcs, facs = "", dropLevels = NULL,
  var = c("ABUNDANCE", "BIOMASS"), rng = NULL, ggtheme = theme_grey(),
  ncol = 2, nrow = 5, showPlots = TRUE)
}
\arguments{
\item{zcs}{- dataframe from call to one of the calcSizeComps... functions}

\item{facs}{- factors to plot by}

\item{dropLevels}{- factor levels to drop from plots}

\item{var}{- variable type to plot (abundance or biomass)}

\item{rng}{- y-axis range (calculated internally if NULL)}

\item{ggtheme}{- ggplot2 theme}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlots}{- flag to show plots immediately}
}
\value{
list of ggplot2 plot objects
}
\description{
Function to plot size comps from survey data as bars.
}
\details{


The time series for each factor level combination is plotted
separately. Distinct levels of each factor can be dropped from the
final plot by seting dropLevels to a list with names corresponding to
factor columns and values being vectors of factor levels to drop.

One plot is created for each distinct level of 'STRATUM'.
}

