% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledSizeComps.calc.R
\name{resampledSizeComps.calc}
\alias{resampledSizeComps.calc}
\title{Resample haul data to bootstrap size compositions by year/stratum from AFSC trawl survey data}
\usage{
resampledSizeComps.calc(
  N = 100,
  tbl_strata,
  tbl_hauls = NULL,
  tbl_indivs = NULL,
  avgHaulsByStation = TRUE,
  byEW166 = FALSE,
  byEBS = TRUE,
  bySex = FALSE,
  byMaturity = FALSE,
  byShellCondition = FALSE,
  cutpts = seq(from = 25, to = 185, by = 5),
  truncate.low = TRUE,
  truncate.high = FALSE,
  verbosity = 0
)
}
\arguments{
\item{N}{: number of resamples/year}

\item{tbl_strata}{: data frame from call to \link{selectStrata.TrawlSurvey} [required]}

\item{tbl_hauls}{: dataframe from call to \link{selectHauls.TrawlSurvey} [required]}

\item{tbl_indivs}{: dataframe from call to \link{selectIndivs.TrawlSurvey} (or crab survey filename)}

\item{avgHaulsByStation}{: flag (T/F) to average hauls by station before calc'ing size comps}

\item{byEW166}{: flag (T/F) to aggregate size comps to EW166}

\item{byEBS}{: flag (T/F) to aggregate size comps to the EBS}

\item{bySex}{: flag (T/F) to calc by sex}

\item{byMaturity}{: flag (T/F) to calc by maturity state}

\item{byShellCondition}{: flag (T/F) to calc by shell condition}

\item{cutpts}{: vector of cutpoints to create size bins from}

\item{truncate.low}{: flag (T/F) to exclude individuals smaller than minSize}

\item{truncate.high}{: flag (T/F) to exclude individuals larger than maxSize}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
data frame with size comps by stratum. Columns are \cr
\itemize{
\item  i - resampling index (i=0 denotes the obsesrved size comps)
\item  YEAR
\item  STRATUM
\item  STRATUM_AREA
\item  SEX
\item  MATURTITY
\item  SHELL_CONDITION
\item  SIZE
\item  numStations
\item  numHauls
\item  numNonZeroHauls
\item  numIndivs
\item  totABUNDANCE = estimated abundance-by-size (by stratum)
\item  totBIOMASS = estimated biomass-by-size (by stratum)
}
}
\description{
This function resamples size compositions by year/stratum for AFSC trawl survey data.
}
\details{
This function performs two-stage resampling (hauls w/in strata, individuals w/in hauls)
to calculate a set of resampled size compositions by year. The size comps with resampling index i=0
are the observed size comps.\cr
\cr
Other notes: \cr
\itemize{
\item{Area is in square nautical miles}
\item{Abundance is in 10^6 indivs}
\item{Biomass   is in 10^3 mt}
}
}
