% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledSizeComps.calcEffN.R
\name{resampledSizeComps.calcEffN}
\alias{resampledSizeComps.calcEffN}
\title{Calculate and plot effective N from resampled size compositions}
\usage{
resampledSizeComps.calcEffN(dfr, byFacs = "", nDefault = 200, nStations = 375)
}
\arguments{
\item{dfr}{- dataframe with resampled size compositions}

\item{byFacs}{- vector of column names for factors other than YEAR and STRATUM}

\item{nDefault}{- assumed effective sample size (for plotting)}

\item{nStations}{- typical number of survey stations (for plotting)}
}
\description{
Function to calculate and plot effective N from resampled size compositions.
}
\details{
dfr should be an output from \code{\link{resampledSizeComps.calc}}.
}
