% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledSizeComps.CalcStats.R
\name{resampledSizeComps.CalcStats}
\alias{resampledSizeComps.CalcStats}
\title{Calculate statistics on resampled size comps}
\usage{
resampledSizeComps.CalcStats(dfrRZCs)
}
\arguments{
\item{dfrRZCs}{- dataframe with resampled size comps from [resampledSizeComps.calc]}
}
\value{
dataframe with statistics from resampled size comps.
}
\description{
Function to calculate statistics on resampled size comps
}
\details{
Mean, median, and 80% confidence intervals are calculated by year,
stratum, sex, maturity state, and shell condition. Stats are calculated on
total abundance and normalized size comps (the latter preceded by "nrm_" in)
the output dataframe.
}
