% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reshapeSizeComps.wide}
\alias{reshapeSizeComps.wide}
\title{Reshape survey size comps dataframe to wide format.}
\usage{
reshapeSizeComps.wide(dfr, facs = "", var = c("totABUNDANCE", "totBIOMASS"),
  export = FALSE, out.csv = "SurveySizeComps.wide.csv", out.dir = NULL,
  verbosity = 1)
}
\arguments{
\item{dfr}{- dataframe with survey size comps from call to \code{\link{calcSizeComps.ByStratum}} or  \code{\link{calcSizeComps.EW166}} or  \code{\link{calcSizeComps.EBS}}}

\item{facs}{- vector of 'factor' column names (e.g., 'SEX', 'MATURITY') for output}

\item{var}{- name of variable to extract ("totABUNDANCE" or "totBIOMASS")}

\item{export}{- boolean flag to write results to csv file}

\item{out.csv}{- output file name}

\item{out.dir}{- output file directory}

\item{verbosity}{- integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
dataframe with size comps in 'wide' format.
}
\description{
Function to reshape a survey size comps dataframe to wide format
}
\details{


Factors (e.g., 'SEX', 'MATURITY', 'SHELL_CONDITION') in the input dataframe that are
missing from the 'facs' vector will be aggregated over by summing. STRATUM is \bold{not} regarded as
a potential factor.

Output column order will be STRATUM, facs, YEAR, numStations, numIndivs, size bins
}

