% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectHauls.TrawlSurvey.R
\name{selectHauls.TrawlSurvey}
\alias{selectHauls.TrawlSurvey}
\title{Function to extract NMFS crab survey hauls data from Excel csv file format by year.}
\usage{
selectHauls.TrawlSurvey(
  tbl_strata,
  tbl = NULL,
  Years = NULL,
  HaulTypes = NULL,
  YearRange = NULL,
  DepthRange = NULL,
  LatRange = NULL,
  LonRange = NULL,
  skip = 6,
  export = FALSE,
  out.csv = "SelectedSurveyHauls.csv",
  out.dir = NULL,
  verbosity = 0
)
}
\arguments{
\item{tbl_strata}{: dataframe from call to \code{\link[=selectStrata.TrawlSurvey]{selectStrata.TrawlSurvey()}} (required)}

\item{tbl}{: trawl survey dataframe from previous call (or name of trawl survey csv datafile, or NULL)              (ignored if NULL)}

\item{Years}{: vector of survey years to include in output             (ignored if NULL)}

\item{HaulTypes}{: vector of haul types to include in output               (ignored if NULL)}

\item{YearRange}{: vector of min, max survey years to include in output    (ignored if NULL)}

\item{DepthRange}{: vector of min, max haul depths to include in output     (ignored if NULL)}

\item{LatRange}{: vector of min, max haul latitudes to include in output  (ignored if NULL)}

\item{LonRange}{: vector of min, max haul longitudes to include in output (ignored if NULL)}

\item{skip}{: number of rows to skip when reading crabhaul csv file (default=6 to match AKFIN crabhaul download)}

\item{export}{: boolean flag to export results to csv file}

\item{out.csv}{: name of output csv file                    (ignored if NULL)}

\item{out.dir}{: base path for output csv file              (set to folder of input csv file or current working directory)}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
a dataframe w/ columns \cr
\itemize{
\item{YEAR}
\item{STRATUM}
\item{GIS_STATION}
\item{HAULJOIN}
\item{HAUL_TYPE}
\item{START_DATE - format: mmddyyyy}
\item{START_HOUR - format: hhmm}
\item{MID_LATITUDE}
\item{MID_LONGITUDE}
\item{BOTTOM_DEPTH - in meters}
\item{GEAR_TEMPERATURE - in deg C}
\item{AREA_SWEPT_VARIABLE - in nm^2}
}
}
\description{
Function to extract NMFS crab survey hauls data from Excel csv file format by year.
}
\details{
If neither tbl or in.csv is given, the user will be prompted for a csv file via a file dialog box
}
