% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.data.R
\docType{data}
\name{shp.EBS.lines.bathy}
\alias{shp.EBS.lines.bathy}
\title{A fortified (spatial) data frame reflecting EBS shelf bathymetry contours (lines).}
\format{A fortified spatial data frame}
\usage{
shp.EBS.lines.bathy
}
\description{
A fortified (spatial) data frame reflecting EBS shelf bathymetry contours (lines). Derived
from the shape file 'data-raw/MapData/Bathymetry/ShelfBathymetry.shp'.
}
\keyword{datasets}
