% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.data.R
\docType{data}
\name{shp.EBS.polys.land}
\alias{shp.EBS.polys.land}
\title{A fortified (spatial) data frame reflecting an Alaska land mask (polygons).}
\format{A fortified spatial data frame}
\usage{
shp.EBS.polys.land
}
\description{
A fortified (spatial) data frame reflecting an Alaska land mask (polygons). Derived
from the shape file 'data-raw/MapData/Land/Alaska.shp'.
}
\keyword{datasets}

