% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.data.R
\docType{data}
\name{shp.EBS.polys.surveyblocks}
\alias{shp.EBS.polys.surveyblocks}
\title{A fortified (spatial) data frame reflecting EBS survey blocks (polygons).}
\format{A fortified spatial data frame}
\usage{
shp.EBS.polys.surveyblocks
}
\description{
A fortified (spatial) data frame reflecting EBS survey blocks (polygons). Derived
from the shape file 'data-raw/MapData/Surveys/NMFS_EBSGroundfishSurveyBlocks.shp'.
}
\keyword{datasets}
