% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamen.RecolorTonerMap.R
\name{stamen.RecolorTonerMap}
\alias{stamen.RecolorTonerMap}
\title{Color land, water in a ggmap object derived from a Stamen toner map.}
\usage{
stamen.RecolorTonerMap(map, land = "khaki", water = "lightblue")
}
\arguments{
\item{map}{- ggmap object derived from a Stamen toner map}

\item{land}{- color specification for land pixels}

\item{water}{- color specification for water pixels}
}
\value{
a ggmap object
}
\description{
Function to color land, water in a ggmap object derived from a Stamen toner map.
}
\details{
Changes land ("#999999) and water ("#000000) pixels in a ggmap object
derived from a Stamen toner map to the specified colors. Attributes of the original
object are copied to the new object.
}

