% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.CreateBaseMap.R
\name{tmap.CreateBaseMap}
\alias{tmap.CreateBaseMap}
\title{Create a basemap layer for maps based on the tmap package}
\usage{
tmap.CreateBaseMap(layer.land = NULL, layer.bathym = NULL,
  layers.survey = NULL, gisDir = system.file("extdata/Shapefiles",
  package = "wtsGIS"),
  shapeFile.bathymetry = "Bathymetry/ShelfBathymetry.shp",
  shapeFile.land = "Land/Alaska.shp", shapeFiles.survey = list(grid =
  "NMFS_Survey_Info/NMFS_EBSSurveyBlocks.shp", stations =
  "NMFS_Survey_Info/NMFS_EBSSurveyStations.PointsLL.shp"),
  strCRS = tmaptools::get_proj4("longlat", output = "character"),
  boundingbox = list(bottomleft = list(lon = -179, lat = 54), topright =
  list(lon = -157, lat = 62.5)), colors.bathym = "darkblue",
  points.size = 0.01)
}
\arguments{
\item{layer.land}{- a tmap layer representing land}

\item{layer.bathym}{-}

\item{layers.survey}{-}

\item{gisDir}{- path to top level folder for shapefiles}

\item{shapeFile.bathymetry}{- bathymetry shapefile}

\item{shapeFile.land}{- land shapefile}

\item{shapeFiles.survey}{- list (grid, stations) of shapfiles to define survey stations}

\item{strCRS}{- string representation of CRS (default = WGS84) used for ALL shapefiles}

\item{boundingbox}{- a bounding box (list with elements "bottomleft" and "topright")}

\item{colors.bathym}{- color for the bathymetry}

\item{points.size}{- size for the station locations}
}
\value{
- basemap layer in WGS based on the tmap package
}
\description{
This function creates a basemap layer for maps based on the tmap package
}
\details{
Uses \code{wtsGIS::tmap.CreateLayerFromShapefile}.
}
