% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.CreateSurveyGridLayers.R
\name{tmap.CreateSurveyGridLayers}
\alias{tmap.CreateSurveyGridLayers}
\title{Create the polygon and point geometry tmap layers from polygon and point shapefiles for a survey grid}
\usage{
tmap.CreateSurveyGridLayers(gisDir = "~", shapeFiles = list(grid =
  "NMFS Survey Info/NMFS_EBSSurveyBlocks.shp", stations =
  "NMFS Survey Info/NMFS_EBSSurveyStations.PointsLL.shp"),
  strCRS = tmaptools::get_proj4("longlat"))
}
\arguments{
\item{gisDir}{- common toplevel directory for files}

\item{shapeFiles}{- a list of shapefiles to read ("grid" and "stations")}

\item{strCRS}{- string describing the final coordinate reference system (CRS)}
}
\value{
a 2-element list with the grid and stations map layers consistent with the tmap package
}
\description{
This function reads shapefile and creates polygon and point tmap layers for a survey grid.
}
\details{
Uses \code{tmaptools::read_shape()} and \code{sp::spTransform()}.
}
