% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_CreateBasemap.R
\name{tmap_CreateBasemap}
\alias{tmap_CreateBasemap}
\title{Create a basemap layer for maps based on the \pkg{tmap} package}
\usage{
tmap_CreateBasemap(
  layer.land = NULL,
  layer.bathym = NULL,
  layers.survey = NULL,
  final.crs = wtsGIS::get_crs("WGS84"),
  bbox = wtsGIS::getStandardBBox("EBS"),
  colors.bathym = "darkblue",
  points.size = 0.01
)
}
\arguments{
\item{layer.land}{- a spatial layer representing land (or NULL)}

\item{layer.bathym}{- a spatial layer representing bathymetry (or NULL)}

\item{layers.survey}{-  a list of spatial layers representing the survey grid and survey stations (or NULL)}

\item{final.crs}{- string representation of CRS (default = WGS84) used for ALL shapefiles}

\item{bbox}{- a \code{sf::bbox} bounding box}

\item{colors.bathym}{- color for the bathymetry}

\item{points.size}{- size for the station locations}
}
\value{
- a \pkg{tmap}-style basemap
}
\description{
This function creates a basemap layer for maps based on the \pkg{tmap} package
}
\details{
Uses \code{wtsGIS::tmap_CreateBasemap} and \code{wtsGIS::transformCRS}.
}
