% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfrf.ConvertFormat2NMFS.R
\name{bsfrf.ConvertFormat2NMFS}
\alias{bsfrf.ConvertFormat2NMFS}
\title{Convert BSFRF data to (almost) NMFS CRABHAUL_DATA format}
\usage{
bsfrf.ConvertFormat2NMFS(
  tbl,
  types = c("SBS", "IDX"),
  sizeCol = "fixed_cw",
  bsfrf_species = "BRD",
  nmfs_species_code = 68560,
  nmfs_species_name = "Bairdi Tanner Crab",
  verbosity = TRUE
)
}
\arguments{
\item{tbl}{: filename for BSFRF  csv file to read or dataframe from reading csv file with \code{\link[=bsfrf.ReadCSV]{bsfrf.ReadCSV()}}}

\item{types}{: BSFRF study types to select ("SBS", "IDX", or both)}

\item{bsfrf_species}{: BSFRF species to extract (default = "BRD")}

\item{nmfs_species_code}{: appropriate NMFS species code (default = 68560 for Bairdi)}

\item{nmfs_species_name}{: appropriate NMFS species name (default = 'Bairdi Tanner Crab')}

\item{verbosity}{: integer flag to print debug info}

\item{size_col}{: column name with crab sizes (default = 'fixed_cw')}
}
\value{
dataframe "almost" in CRABHAUL_DATA format.
}
\description{
Function to convert BSFRF data to (almost) NMFS CRABHAUL_DATA format.
}
\details{
BSFRF station ids are not necessarily correct. For example, the
so-called corner stations in the NMFS survey grid use the format 'aaxxxx' whereas
the BSFRF format is typically 'aa-xxxx' , where aa is a 2-letter code and
xxxx is a 4-digit code. NMFS station 'H-21' has also been mis-coded 'H21' in
the BSFRF data in some years. Rather than deal with these instances piecemeal,
the NMFS survey grid layers (\code{\link[=gisGetSurveyGridLayers]{gisGetSurveyGridLayers()}}) are used to reassign NMFS station ids to the
BSFRF data based on the mid-tow coordinates for the BSFRF data. The hauls
that will be changed can be identified by running \code{\link[=bsfrf.CheckStation]{bsfrf.CheckStation()}} on the
BSFRF haul dataframe/csv file before making the conversion here.

returned dataframe has columns
\itemize{
\item {AKFIN_SURVEY_YEAR}
\item {HAULJOIN}
\item {VESSEL}
\item {CRUISE}
\item {HAUL}
\item {HAUL_TYPE}
\item {START_DATE}
\item {AKFIN_SURVEY_DATE}
\item {START_HOUR}
\item {MID_LATITUDE}
\item {MID_LONGITUDE}
\item {GIS_STATION}
\item {BOTTOM_DEPTH (in meters)}
\item {GEAR_TEMPERATURE (deg C)}
\item {AREA_SWEPT_VARIABLE (nm^2)}
\item {SPECIES_CODE}
\item {SPECIES_NAME}
\item {SEX}
\item {WIDTH (mm)}
\item {SHELL_CONDITION}
\item {EGG_COLOR}
\item {EGG_CONDITION}
\item {CLUTCH_SIZE}
\item {CHELA_HEIGHT}
\item {CALCULATED_WEIGHT}
\item {WEIGHT}
\item {SAMPLING_FACTOR}
\item {CPUE_NUM}
}
}
