% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcBiomass.ByStratum}
\alias{calcBiomass.ByStratum}
\title{Calculate average cpue, numbers and biomass by stratum from AFSC trawl survey data.}
\usage{
calcBiomass.ByStratum(tbl_strata, tbl_cpue = NULL, export = FALSE,
  out.csv = "SurveyBiomass.ByStratum.csv", out.dir = NULL, verbosity = 1)
}
\arguments{
\item{tbl_strata}{: data frame w/ stations/strata from call to selectStrata.TrawlSurvey(...)}

\item{tbl_cpue}{: data frame w/ cpue by year, station, other factor levels (or csv filename or NULL)}

\item{export}{: boolean flag to write results to csv file}

\item{out.csv}{: output file name}

\item{out.dir}{: output file directory}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
data frame with average cpue (numbers, weight), abundance and biomass by stratum. Columns are \cr
\itemize{
\item  YEAR
\item  STRATUM
\item  STRATUM_AREA
\item  other user-defined factors
\item  numStations
\item  numHauls
\item  numNonZeroHauls
\item  numIndivs
\item  totABUNDANCE = estimated abundance (by stratum)
\item  stdABUNDANCE = std deviation of estimated abundance (by stratum)
\item  cvABUNDANCE  = cv of estimated abundance (by stratum)
\item  totBIOMASS = estimated biomass (by stratum)
\item  stdBIOMASS = std deviation of estimated biomass (by stratum)
\item  cvBIOMASS  = cv of estimated biomass (by stratum)
}
}
\description{
This function calculates average cpue, numbers and biomass by stratum from cpue (by survey station or by haul).
}
\details{


If tbl_cpue is NULL, the user is prompted to enter a csv file with cpue info. \cr
\cr Other notes: \cr
\itemize{
  \item Area is in square nautical miles
  \item CPUE in numbers is in no/(sq. nm.)
  \item CPUE in weight  is in mt/(sq. nm.)
  \item Abundance is in 10^6 indivs
  \item Biomass   is in 10^3 mt
}
}

