% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiomass.EW166.R
\name{calcBiomass.EW166}
\alias{calcBiomass.EW166}
\title{Calculate abundance, biomass by year, east/west of 166W, and other factors from a biomass-by-stratum data frame or csv file.}
\usage{
calcBiomass.EW166(tbl = NULL,
  strata_toEW166 = Codes.TrawlSurvey()[["strata.EW166"]], export = TRUE,
  out.csv = "SurveyBiomass.EW166.csv", out.dir = NULL, verbosity = 1)
}
\arguments{
\item{tbl}{: data frame with biomass by stratum info from call to \code{\link{calcBiomass.ByStratum}} or csv file with biomass by stratum info, or NULL}

\item{strata_toEW166}{: data frame w/ conversion from original strata to EW166 strata}

\item{export}{: boolean flag to write results to csv file}

\item{out.csv}{: output file name}

\item{out.dir}{: output file directory}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
Dataframe w/ estimates of abundance, biomass by year, strata as east/west of 166W, and other factors. Columns are \cr
\itemize{
\item  YEAR            = survey year
\item  STRATUM         = 'EAST' or 'WEST' of 166W
\item  STRATUM_AREA    = area of stratum
\item  other user-defined factors (e.g., sex, shell_condition)
\item  numStations     = number of stations included
\item  numHauls        = number of hauls included
\item  numNonZeroHauls = number of hauls included
\item  numIndivs       = number of individuals sampled
\item  totABUNDANCE    = estimated abundance
\item  stdABUNDANCE    = std deviation of estimated abundance
\item  cvABUNDANCE     = cv of estiamted abundance
\item  totBIOMASS      = estimated biomass
\item  stdBIOMASS      = std deviation of estimated biomass 
\item  cvBIOMASS       = cv of estimated biomass
}
}
\description{
This function estimates abundance, biomass by year, east/west of 166W from a biomass-by-stratum data frame or csv file.
}
\details{
Note: if tbl and in.csv are both NULL, the user is prompted to enter a csv file with biomass by stratum info. \cr
Notes: \cr
\itemize{
\item   Area is in square nautical miles
\item   Abundance is in 10^6 individuals
\item   Biomass   is in 10^3 mt
}
}

