% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEffectiveAreaOccupied.R
\name{calcEffectiveAreaOccupied}
\alias{calcEffectiveAreaOccupied}
\title{Calculate the effective area occupied by different life stages}
\usage{
calcEffectiveAreaOccupied(dfrCPUE)
}
\arguments{
\item{dfrCPUE}{\itemize{
\item dataframe with cpue estimates by station (i.e., output from from \link{calcCPUE.ByStation})
}}
}
\value{
a 2-element list with dataframes giving effective area occupied based on abundance and biomass.
}
\description{
This function calculate the effective area occupied by different life stages.
}
\details{
Uses the respective survey station area from the survey grid
}
