% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.TrawlSurvey.R
\name{Codes.TrawlSurvey}
\alias{Codes.TrawlSurvey}
\title{Provides a list of dataframes suitable for converting from survey codes to
assessment codes and descriptions.}
\usage{
Codes.TrawlSurvey()
}
\value{
a list
}
\description{
This function provides a list of dataframes suitable for converting
from survey codes to assessment codes and descriptions.
}
\details{
a list with the following elements: \cr
\itemize{
\item sex             : table to convert between survey and assessment sex codes
\item shell_condition : table to convert between survey and assessment shell condition codes
\item clutch_size     : table to convert between survey and assessment clutch size codes
\item strata.orig.BTC : original strata for Bairdi Tanner crab
\item strata.revd.BTC : revised strata for Bairdi Tanner crab
\item strata.2015.BTC : 2015 strata for Bairdi Tanner crab
\item strata.orig.BKC : original strata for blue king crab
\item strata.2015.BKC : 2015 strata for blue king crab
\item strata.2015.OTC : 2015 strata for Opilio Tanner crab (i.e., snow crab)
\item strata.2015.RKC : 2015 strata for red king crab
\item strata.EW166    : E/W 166W strata
}
}
