% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Codes.TrawlSurvey}
\alias{Codes.TrawlSurvey}
\title{Provides a list of dataframes suitable forconverting from survey codes to
assessment codes and descriptions.}
\usage{
Codes.TrawlSurvey()
}
\value{
a list with the following elements: \cr
\itemize{
\item sex             : table to convert between survey and assessment sex codes
\item shell_condition : table to convert between survey and assessment shell condition codes
\item clutch_size     : table to convert between survey and assessment clutch size codes
\item strata.BTC      : revised strata for Bairdi Tanner crab
\item strata.orig.BTC : original strata for Bairdi Tanner crab
\item strata.EW166    : E/W 166W strata
}
}
\description{
This function provides a list of dataframes suitable for converting
from survey codes to assessment codes and descriptions
}
\details{

}

