% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertImmOS.R
\name{convertImmOS}
\alias{convertImmOS}
\title{Re-classify immature, old shell crab as immature, new shell crab}
\usage{
convertImmOS(dfr, colMat = "MATURITY", colSC = "SHELL_CONDITION")
}
\arguments{
\item{dfr}{- dataframe to run re-classification on}

\item{colMat}{- name of column with maturity designation}

\item{colSC}{- name of column with shell condition designation}
}
\value{
dataframe with same strucuture as input and shell condition converted to "NEW_SHELL" for
immature crab formerly classified as "OLD_SHELL" (irrespective of case).
}
\description{
Function to re-classify immature, old shell crab as immature, new shell crab.
}
\details{
None.
}
