% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputFile.SurveyData.R
\name{createInputFile.SurveyData}
\alias{createInputFile.SurveyData}
\title{Creates a survey input file for TCSAM02}
\usage{
createInputFile.SurveyData(fn, survey_name = "NMFS_trawl_survey",
  acdInfo = list(abundance = list(fitType = "BY_XM", likeType =
  "LOGNORMAL", likeWgt = 0), biomass = list(fitType = "BY_X_MAT_ONLY",
  likeType = "LOGNORMAL", likeWgt = 1)), zcsInfo = list(fitType =
  "BY_XME", likeType = "MULTINOMIAL", likeWgt = 1), cutpts, dfrACD, dfrZCs,
  dfrSSs, ssCol = "numIndivs", verbose = FALSE)
}
\arguments{
\item{fn}{- output file name}

\item{survey_name}{- name to use for survey}

\item{acdInfo}{- list with sublists for abundance and biomass specifying fitType, likeType, and likeWgt}

\item{zcsInfo}{- list for size compositions specifying fitType, likeType, and likeWgt}

\item{cutpts}{- vector of cutpoints for the input size compositions}

\item{dfrACD}{- dataframe of aggregated catch data}

\item{dfrZCs}{- dataframe of size comps}

\item{dfrSSs}{- dataframe with relative sample sizes to use with size comps}

\item{verbose}{- flag to print debugging info}

\item{colSS}{- sample size column name in dfrSSs (default="numIndivs")}
}
\description{
This function that creates a survey input file for TCSAM02.
}
\details{
This function that creates a survey input file for TCSAM02.
}
