% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCalcs_CPUEbyStation.R
\name{doCalcs_CPUEbyStation}
\alias{doCalcs_CPUEbyStation}
\title{Calculate cpue by survey station and other factors from strata and haul files}
\usage{
doCalcs_CPUEbyStation(tbl_strata, tbl_hauls, tbl_indivs,
  aggBySex = FALSE, aggByMaturity = FALSE,
  aggByShellCondition = FALSE, aggBySize = FALSE, cutpts = seq(from =
  25, to = 185, by = 5), truncate.low = TRUE, truncate.high = FALSE,
  dropLevels = list(SEX = c("MISSING", "HERMAPHRODITIC")),
  verbosity = 0)
}
\arguments{
\item{tbl_strata}{- dataframe with survey strata info (output from \code{\link{selectStrata.TrawlSurvey}})}

\item{tbl_hauls}{- dataframe with hauls info (output from \code{\link{selectHauls.TrawlSurvey}})}

\item{tbl_indivs}{- dataframe with individual crab info (output from \code{\link{selectIndivs.TrawlSurvey}})}

\item{aggBySex}{- flag to agregate CPUE over sexes}

\item{aggByMaturity}{- flag to aggregate of maturity states}

\item{aggByShellCondition}{- flag to aggregate over shell conditions}

\item{aggBySize}{- flag to aggregate over sizes}

\item{cutpts}{- vector of cutpoints to create size bins from}

\item{truncate.low}{- flag (T/F) to exclude individuals smaller than minSize}

\item{truncate.high}{- flag (T/F) to exclude individuals larger than maxSize}

\item{dropLevels}{- NULL, or list (by factor name) of vectors of factor values to drop (see help for \code{wtsUtilities::dropLevels})}

\item{verbosity}{- integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
A dataframe of cpue (numbers and weight) by survey station. See Details.
}
\description{
Function to calculate cpue by survey sstation and other factors (e.g., sex) from strata and haul files.
}
\details{
The returned dataframes has the following columns: \cr
\itemize{
\item   YEAR
\item   STRATUM
\item   GIS_STATION
\item   LONGITUDE
\item   LATITUDE
\item   SEX
\item   MATURITY
\item   SHELL_CONDITION
\item   SIZE
\item   numHauls
\item   numNonZeroHauls
\item   numIndivs
\item   numCPUE
\item   wgtCPUE
}
\cr Other notes: \cr
\itemize{
\item   CPUE in numbers is in no/(sq. nm.)
\item   CPUE in weight  is in mt/(sq. nm.)
} \cr
}
