% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCalcs_ZCs.R
\name{doCalcs_ZCs}
\alias{doCalcs_ZCs}
\title{Function to calculate size compositions from station, haul and individual crab info.}
\usage{
doCalcs_ZCs(
  tbl_strata,
  tbl_hauls,
  tbl_indivs,
  useStratumArea = TRUE,
  calcByEW166 = TRUE,
  aggBySex = FALSE,
  aggByMaturity = FALSE,
  aggByShellCondition = FALSE,
  cutpts = seq(from = 0, to = 185, by = 5),
  truncate.low = TRUE,
  truncate.high = FALSE,
  recodeOldShellImmature = FALSE,
  dropLevels = list(SEX = c("MISSING", "HERMAPHRODITIC")),
  verbosity = 0
)
}
\arguments{
\item{tbl_strata}{: dataframe with survey strata info (output from \code{\link{selectStrata.TrawlSurvey}})}

\item{tbl_hauls}{: dataframe with hauls info (output from \code{\link{selectHauls.TrawlSurvey}})}

\item{tbl_indivs}{: dataframe with individual crab info (output from \code{\link{selectIndivs.TrawlSurvey}})}

\item{useStratumArea}{: flag (T/F) to use STRATUM_AREA to expand average CPUE to stratum abundance/biomass (default is T)}

\item{aggBySex}{\itemize{
\item flag to agregate CPUE over sexes
}}

\item{aggByMaturity}{\itemize{
\item flag to aggregate of maturity states
}}

\item{aggByShellCondition}{\itemize{
\item flag to aggregate over shell conditions
}}

\item{cutpts}{: seq(from=0,to=185,by=5)}

\item{truncate.low}{: TRUE}

\item{truncate.high}{: FALSE}

\item{recodeOldShellImmature}{: recode old shell, immature crab as new shell immature crab}

\item{dropLevels}{\itemize{
\item NULL, or list (by factor name) of vectors of factor values to drop (see help for \code{wtsUtilities::dropLevels})
}}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}

\item{calcByEW166:}{TRUE (calculate size compositions by EW166 prior to calculating total for EBS)}
}
\value{
a list with the following elements: \cr
\itemize{
\item {EBS}        {- data frame of size compositions by year, sex, maturity state, and shell condition}
\item {EW166}      {- data frame of size compositions by EW166, year, sex, maturity state, and shell condition}
\item {byStratum}  {- data frame of size compositions by survey stratum, year, sex, maturity state, and shell condition}
}
}
\description{
Function to calculate size compositions from station, haul and individual crab info.
}
\details{
Notes: \itemize{
\item abundance is in 10^6 individuals
}
}
