% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSizeComps.wide.R
\name{exportSizeComps.wide}
\alias{exportSizeComps.wide}
\title{Export survey size comps dataframe in wide format for the stock assessment.}
\usage{
exportSizeComps.wide(
  dfr,
  facs = "",
  var = c("ABUNDANCE", "BIOMASS"),
  dropLevels = NULL,
  export = FALSE,
  out.csv = paste("SurveySizeComps", var[1], "wide.csv", sep = "."),
  out.dir = NULL,
  verbosity = 0
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe with survey size comps from call to \code{\link{calcSizeComps.ByStratum}} or  \code{\link{calcSizeComps.EW166}} or  \code{\link{calcSizeComps.EBS}}
}}

\item{facs}{\itemize{
\item vector of 'factor' column names (e.g., 'SEX', 'MATURITY') for output
}}

\item{var}{\itemize{
\item variable type to extract ("ABUNDANCE" or "BIOMASS")
}}

\item{dropLevels}{\itemize{
\item factor levels to drop
}}

\item{export}{\itemize{
\item boolean flag to write results to csv file
}}

\item{out.csv}{\itemize{
\item output file name
}}

\item{out.dir}{\itemize{
\item output file directory
}}

\item{verbosity}{\itemize{
\item integer flag indicating level of printed output (0=off,1=minimal,2=full)
}}
}
\value{
dataframe with size comps in 'wide' format.
}
\description{
Function to reshape a survey size comps dataframe in wide format for the stock assessment.
}
\details{
Factors (e.g., 'SEX', 'MATURITY', 'SHELL_CONDITION') in the input dataframe that are
missing from the 'facs' vector will be aggregated over by summing. STRATUM is \bold{not} regarded as
a potential factor.

Output column order will be STRATUM, facs, YEAR, numStations, numIndivs, size bins
}
