% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getChelaHeights.R
\name{getChelaHeights}
\alias{getChelaHeights}
\title{Function to extract chela height data on males w/ selected characteristics from NMFS trawl survey csv files.}
\usage{
getChelaHeights(tbl_hauls, tbl = NULL, col.Size = "WIDTH",
  shell_condition = c("NEW_SHELL", "OLD_SHELL", "ALL"), minSize = -Inf,
  maxSize = Inf, showPlot = TRUE, export = FALSE,
  out.csv = "ChelaHeights.csv", out.dir = NULL, verbosity = 1, ...)
}
\arguments{
\item{tbl_hauls}{- hauls table (dataframe) from call to selectHauls.TrawlSurvey(...) [required]}

\item{tbl}{- table (dataframe) of survey data (or filename of survey csv file, or NULL)}

\item{shell_condition}{- one of 'NEW_SHELL','OLD_SHELL' or 'ALL' for narrowing selection of individuals}

\item{minSize}{- minimum size (width) of individuals to select}

\item{maxSize}{- maximum size (width) of individuals to select}

\item{export}{- boolean flag to export results to csv file}

\item{out.csv}{- name of output csv file                    (ignored if NULL)}

\item{out.dir}{- base path for output csv file              (set to folder of input csv file or current working directory)}

\item{verbosity}{- flags for intermediate output}

\item{...:}{parameters passed to plotChelaHeights(...)}
}
\value{
dataframe
}
\description{

}
\details{
If tbl is NULL, the user will be prompted for a survey csv file via a file dialog box.\cr
\itemize{\item Weights are in grams.}
}

