% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_MapCPUE.R
\name{gg_MapCPUE}
\alias{gg_MapCPUE}
\title{Create a \pkg{ggplot2} map of cpue-by-haul or by-station}
\usage{
gg_MapCPUE(
  dfrCPUE,
  type = c("biomass", "abundance"),
  column = NULL,
  minVal = 0,
  maxVal = Inf,
  label = ifelse(!is.null(column), column, type[1]),
  basemap = gg_GetBasemapLayers(),
  plotAsPolys = TRUE,
  plotLand = TRUE,
  plotBathymetry = TRUE,
  plotSurveyGrid = TRUE,
  color_scale = ggplot2::scale_colour_viridis_c(name = label, option = "plasma"),
  fill_scale = ggplot2::scale_fill_viridis_c(name = label, option = "plasma"),
  size_scale = ggplot2::scale_size_area(name = label),
  facet_grid = NULL,
  facet_wrap = NULL,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed"
)
}
\arguments{
\item{dfrCPUE}{- dataframe with cpue-by-haul or by-station (see \link{@details})}

\item{type}{- string indicating cpue type: "biomass" or "abundance" (or "other")}

\item{column}{- name of column (if not wgtCPUE or numCPUE, which are identified automatically)}

\item{minVal}{- min value for scales (values <= minVal are not plotted; default = 0)}

\item{maxVal}{- max value for scales (values > maxVal will be set to maxVal; default = Inf)}

\item{label}{- label for color/fill/size legends}

\item{basemap}{- \pkg{ggplot2} layers for basemap of plots (see [gg_GetBasemapLayers()])}

\item{plotAsPolys}{- flag (T/F) to plot as cpue as colored polygons or as colored, scaled points}

\item{plotLand}{- flag (T/F) to include land in plot}

\item{plotSurveyGrid}{- flag (T/F) to include survey grid in plot}

\item{color_scale}{- \pkg{ggplot2} scale_color object to use for colors}

\item{fill_scale}{- \pkg{ggplot2} scale_fill object to use for fills}

\item{size_scale}{- \pkg{ggplot2} scale_size object to use for symbol sizes}

\item{facet_grid}{- formula to use for gridding plot facets}

\item{facet_wrap}{- formula to use for wrapping plot facets}

\item{nrow}{- number of rows for wrapped facets}

\item{ncol}{- number of columns for wrapped facets}

\item{scales}{- string indicating how to treat scales among facets}

\item{plotBathym}{- flag (T/F) to include bathymetry in plot}
}
\value{
a \pkg{ggplot2} plot object.
}
\description{
Function to create a \pkg{ggplot2} map of survey cpue-by-haul or by-station.
}
\details{
If dfrCPUE is not a \pkg{sf} dataframe, then the function calls
[getSurveyGridLayers()] to create a pkg{sf} dataframe using [wtsGIS::mergeDataframeWithLayers()],
merging the CPUE dataframe with either the \code{grid} or \code{points} element of the survey
grid layers list. The dataframe is joined to the gis layer using the `GIS_STATION` and `ID`
columns. If `column` is NULL, then it is set to "wgtCPUE" if \code{type="biomass"}
or "numCPUE" if \code{type="abundance"}.

An alternative is provide a \pkg{sf} dataframe, in which case data in the column identified by
`column` is mapped using the active geometry (in this case `plotAsPolys` and `plotAsPoints` are ignored).
This provides a way of "faking" the code to plot any scalar data associated with a \pkg{sf} dataframe.

#'
}
