% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_MapCPUE.R
\name{gg_MapCPUE}
\alias{gg_MapCPUE}
\title{Create a \pkg{ggplot2} map of cpue-by-haul or by-station}
\usage{
gg_MapCPUE(
  dfrCPUE,
  type = c("biomass", "abundance"),
  basemap = gg_GetBasemapLayers(),
  plotAsPolys = TRUE,
  color_scale = ggplot2::scale_colour_viridis_c(option = "plasma"),
  fill_scale = ggplot2::scale_fill_viridis_c(option = "plasma"),
  size_scale = ggplot2::scale_size_area(),
  facet_grid = NULL,
  facet_wrap = NULL,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed"
)
}
\arguments{
\item{dfrCPUE}{- dataframe with cpue-by-haul or by-station}

\item{type}{- string indicating cpue type: "biomass" or "abundance"}

\item{basemap}{- \pkg{ggplot2} layers for basemap of plots (see \code{\link{gg_GetBasemapLayers}})}

\item{plotAsPolys}{- flag (T/F) to plot as cpue as colored polygons or as colored, scaled points}

\item{color_scale}{- \pkg{ggplot2} scale_color object to use for colors}

\item{fill_scale}{- \pkg{ggplot2} scale_fill object to use for fills}

\item{size_scale}{- \pkg{ggplot2} scale_size object to use for symbol sizes}

\item{facet_grid}{- formula to use for gridding plot facets}

\item{facet_wrap}{- formula to use for wrapping plot facets}

\item{nrow}{- number of rows for wrapped facets}

\item{ncol}{- number of columns for wrapped facets}

\item{scales}{- string indicating how to treat scales among facets}
}
\value{
\pkg{ggplot2} plot object
}
\description{
Function to create a \pkg{ggplot2} map of survey cpue-by-haul or by-station
}
\details{
None.
}
