% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisCreateSurveyGridLayers.R
\name{gisCreateSurveyGridLayers}
\alias{gisCreateSurveyGridLayers}
\title{Create \code{sf} dataframes for a survey grid from polygon and point shapefiles}
\usage{
gisCreateSurveyGridLayers(
  gisPath = NULL,
  shapeFiles = list(grid = NULL, stations = NULL),
  final_crs = wtsGIS::get_crs("WGS84")
)
}
\arguments{
\item{gisPath}{- path to common toplevel folder for files}

\item{shapeFiles}{- a list of shapefiles to read ("grid" and "stations", either can be NULL to skip creation of that layer)}

\item{final_crs}{- a \code{sf::crs} object to be used as the final coordinate reference system for the spatial layers}
}
\value{
a 2-element list with the grid and stations spatial layers as \pkg{sf} dataframes.
}
\description{
This function reads shapefile and creates \pkg{sf} polygon and point dataframes for a survey grid.
}
\details{
None.
}
