% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisCreateSurveyGridLayers.R
\name{gisCreateSurveyGridLayers}
\alias{gisCreateSurveyGridLayers}
\title{Create the polygon and point geometry tmap layers from polygon and point shapefiles for a survey grid}
\usage{
gisCreateSurveyGridLayers(
  gisPath = NULL,
  shapeFiles = list(grid = NULL, stations = NULL),
  strCRS = wtsGIS::getCRS("WGS84"),
  as.sf = TRUE
)
}
\arguments{
\item{gisPath}{- path to common toplevel folder for files}

\item{shapeFiles}{- a list of shapefiles to read ("grid" and "stations"), or NULL to use defaults for the NMFS EBS survey}

\item{strCRS}{- string describing the final coordinate reference system (CRS) (or NULL to use default)}

\item{as.sf}{- flag to create grid layers using simple features (sf) datasets (TRUE) or sp datasets (FALSE)}
}
\value{
a 2-element list with the grid and stations map layers consistent with the tmap package
}
\description{
This function reads shapefile and creates polygon and point tmap layers for a survey grid.
}
\details{
Uses \code{tmaptools::read_shape()} and \code{sp::spTransform()}.
}
