% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisGetSurveyGridLayers.R
\name{gisGetSurveyGridLayers}
\alias{gisGetSurveyGridLayers}
\title{Retrieve the polygon and point geometry tmap layers from polygon and point shapefiles for a survey grid}
\usage{
gisGetSurveyGridLayers()
}
\value{
a 2-element list with the grid and stations map layers consistent with the tmap package
}
\description{
This function retrieves the polygon and point tmap layers for a survey grid.
}
\details{
None.
}
