% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotACD.R
\name{plotACD}
\alias{plotACD}
\title{Plot time series of aggregated catch data.}
\usage{
plotACD(acd, ylab = "", ci = 0.8, ci.type = "normal", xlims = NULL,
  ylims = NULL, ggtheme = theme_grey(), jitter = TRUE, ncol = 2,
  nrow = 5, showPlots = TRUE, verbosity = 0)
}
\arguments{
\item{acd}{- dataframe from call to aggregateCatchData(...)}

\item{ylab}{- y-axis label}

\item{ci}{- confidence interval for error bars (e.g., 0.95)}

\item{ci.type}{- confidence interval type ('normal', 'lognormal')}

\item{xlims}{- x-axis limits (calculated internally if NULL)}

\item{ylims}{- y-axis limits (calculated internally if NULL)}

\item{ggtheme}{- ggplot2 theme}

\item{jitter}{- flag to jitter points}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlots}{- flag to show plots immediately}

\item{verbosity}{- flag (>0) to print diagnostic output}
}
\value{
list of ggplot2 plot objects
}
\description{
Function to plot time series of aggregated catch data.
}
\details{
The time series for each factor level combination is plotted
separately. Distinct levels of each factor can be dropped from the
final plot by seting dropLevels to a list with names corresponding to 
factor columns and values being vectors of factor levels to drop.

One plot is created for each distinct level of 'STRATUM'.
}

