% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAZC.R
\name{plotAZC}
\alias{plotAZC}
\title{Plot aggregated size comps from survey data as bars.}
\usage{
plotAZC(zcs, byStratum = TRUE, multipliers = NULL,
  zlab = "Abundance (millions)", zlims = NULL, xlims = NULL,
  ylims = NULL, ggtheme = theme_grey(), ncol = 2, nrow = 5,
  showPlots = TRUE, verbosity = 0)
}
\arguments{
\item{zcs}{- dataframe from call to aggregateSizeComps(...)}

\item{byStratum}{- flag to plot size comps separately by stratum}

\item{multipliers}{- list by factor of multipliers for factor levels}

\item{zlab}{- 'z'-axis label}

\item{zlims}{- z-axis range (calculated internally if NULL)}

\item{xlims}{- size limits (calculated internally if NULL)}

\item{ylims}{- year limits (calculated internally if NULL)}

\item{ggtheme}{- ggplot2 theme}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlots}{- flag to show plots immediately}

\item{verbosity}{- integer flag (>0) to print diagnostic output}
}
\value{
list of ggplot2 plot objects
}
\description{
Function to plot aggregated size comps from survey data as bars.
}
\details{
The time series for each factor level combination is plotted
separately. Distinct levels of each factor can be dropped from the
final plot by seting dropLevels to a list with names corresponding to 
factor columns and values being vectors of factor levels to drop.

'multipliers' is a list of lists, with each sublist has two elements, 'factors' and 'value'.
'factors' should be a list of factor=level pairs. 'value' should be the value to multiply by. 

One plot is created for each distinct level of 'STRATUM'.
}

