% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChelaHeights.R
\name{plotChelaHeights}
\alias{plotChelaHeights}
\title{Plot chela height data}
\usage{
plotChelaHeights(dfrp, ratio = 0.18, discrim = NULL, useLnScales = FALSE,
  aggYears = 5, rng = NULL, ncol = 2, showPlot = FALSE)
}
\arguments{
\item{dfrp}{- dataframe of chela height data from call to getChelaHeights(...)}

\item{ratio}{- default ratio}

\item{discrim}{- optional function taking a vector of sizes (CW) and returning the associated chela height cutline values yielding a "mature" classification}

\item{useLnScales}{- flag (T/F) to plot data on ln-scales}

\item{aggYears}{- number of years over which to aggregate}

\item{rng}{- y-axis range}

\item{ncol}{- number of coulmns to use for faceting}

\item{showPlot}{- flag to show plots immediately}
}
\value{
ggplot2 plot object to print
}
\description{
Function to plot chela height data
}
\details{
none.
}
