% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGGMaps.CPUE.R
\name{plotGGMaps.CPUE}
\alias{plotGGMaps.CPUE}
\title{Plot maps of survey cpue by year and other factors using ggmap and ggplot2.}
\usage{
plotGGMaps.CPUE(cpue, type = c("abundance", "biomass"), facs = c("SEX",
  "MATURITY", "SHELL_CONDITION"), keepLevels = list(SEX = c("MALE", "FEMALE"),
  MATURITY = c("IMMATURE", "MATURE"), SHELL_CONDITION = c("NEW_SHELL",
  "OLD_SHELL")), dropLevels = list(SEX = c("HERMAPHRODITE", "MISSING"),
  MATURITY = c("MISSING"), SHELL_CONDITION = c("MISSING")), max = NULL,
  scale = 1, map = getMapData("ggmap", "EBS.stamen.toner.z06"),
  bbox = c(left = -180, bottom = 54, right = -155, top = 63),
  mapcolors = list(land = "darkgreen", water = "lightblue"),
  bathymetry = getMapData("shape", "EBS.lines.bathy"),
  land = getMapData("shape", "EBS.polys.land"), layers = NULL, label = "",
  ggtheme = theme_grey(), ncol = 1, nrow = 4, showPlot = FALSE,
  verbosity = 0)
}
\arguments{
\item{cpue}{- dataframe of cpue from call to calcCPUE.ByHaul or calcCPUE.ByStation}

\item{type}{- type of cpue to plot ('abundance' or 'biomass')}

\item{facs}{- vector of factor column names to plot by (e.g., c('SEX', 'MATURITY')) or c() to ignore.}

\item{keepLevels}{- list (by elements of facs) of factor levels to keep before plotting}

\item{dropLevels}{- list (by elements of facs) of factor levels to drop before plotting}

\item{max}{- max for ALL factor combinations (NULL -> scale by max w/in each factor level combination)}

\item{scale}{- value to scale cpue by before plotting (max should be expressed in scale units)}

\item{map}{- ggmap raster object for base map background}

\item{bbox}{- coordinates of bounding box for map (left, bottom, right, top)}

\item{mapcolors}{- list of colors (land, water) for base map background}

\item{bathymetry}{- fortified spatial dataframe or list(dsn=,layer=) specifying bathymetry shapefile to include on base map}

\item{land}{- fortified spatial dataframe or list(dsn=,layer=) specifying land shapefile to include on base map}

\item{layers}{- list of ggplot layers to include in base map}

\item{label}{- page title, if 'facs' is empty (length 0)}

\item{ggtheme}{- ggplot2 theme}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlot}{- flag to show plots immediately}

\item{verbosity}{- integer flag (>0) to print intermediate info}
}
\value{
list of ggplot2 plot objects to print
}
\description{
Function to plot maps of survey cpue by year and other factors using ggmap and ggplot2.
}
\details{
none.
}

