% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotGGMaps.CPUE.R
\name{plotGGMaps.CPUE}
\alias{plotGGMaps.CPUE}
\title{Plot maps of survey cpue by year and other factors using ggmap and ggplot2.}
\usage{
plotGGMaps.CPUE(cpue, type = c("abundance", "biomass"), facs = c("SEX",
  "MATURITY", "SHELL_CONDITION"), keepLevels = list(SEX = c("MALE", "FEMALE"),
  MATURITY = c("IMMATURE", "MATURE"), SHELL_CONDITION = c("NEW_SHELL",
  "OLD_SHELL")), dropLevels = list(SEX = c("HERMAPHRODITE", "MISSING"),
  MATURITY = c("MISSING"), SHELL_CONDITION = c("MISSING")), max = NULL,
  bbox = c(left = -180, bottom = 54, right = -155, top = 63),
  bathymetry = list(dsn = path.expand("~/MapData/Bathymetry"), layer =
  "ShelfBathymetry"), land = list(dsn = path.expand("~/MapData/Land"), layer =
  "Alaska"), layers = NULL, ggtheme = theme_grey(), ncol = 1, nrow = 4,
  showPlots = FALSE)
}
\arguments{
\item{cpue}{- dataframe of cpue from call to calcCPUE.ByHaul or calcCPUE.ByStation}

\item{type}{- type of cpue to plot ('abundance' or 'biomass')}

\item{facs}{- factors to plot (e.g., 'SEX', 'MATURITY')}

\item{keepLevels}{- list (by elements of facs) of factor levels to keep before plotting}

\item{dropLevels}{- list (by elements of facs) of factor levels to drop before plotting}

\item{max}{- max for ALL factor combinations (NULL -> scale by max w/in each factor level combination)}

\item{bbox}{- coordinates of bounding box for map (left, bottom, right, top)}

\item{bathymetry}{- list(dsn=,layer=) specifying bathymetry shapefile to include on map}

\item{land}{- list(dsn=,layer=) specifying land shapefile to include on map}

\item{layers}{- list of lists(dsn=,layer=) specifying other shapefiles to include on map}

\item{ggtheme}{- ggplot2 theme}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlots}{- flag to show plots immediately}
}
\value{
list of ggplot2 plot objects to print
}
\description{
Function to plot maps of survey cpue by year and other factors using ggmap and ggplot2.
}
\details{

}

