% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMaps.CPUE.R
\name{plotMaps.CPUE}
\alias{plotMaps.CPUE}
\title{Plot spatial maps of cpue by year, survey haul or station and other factors from cpue info.}
\usage{
plotMaps.CPUE(tbl_cpue, years = NULL, ztype = "numCPUE", scaleBy = 1,
  title = "Survey", zlab = "no. crab", zunits = "num/nm@+2@+",
  zscl = NULL, xyrng = "180/205/54/62", rotate = 170, elev = 45,
  delx = 0.5, dely = 0.25, blocktype = c("MEAN", "SUM"),
  plt_blocktype = c("SMOOTH", "COARSE"), logtr = FALSE, plt_title = FALSE,
  plt_bars = TRUE, plt_surface = FALSE, plt_blocklocations = plt_surface,
  plt_colorscale = plt_surface | plt_bars, plt_stations = TRUE,
  plt_reflines = TRUE, reflines = list(list(lon = -166 + 0 * seq(from = 50,
  to = 80, by = 1), lat = seq(from = 50, to = 80, by = 1))),
  bathymetryFile = "/Users/WilliamStockhausen/Programming/R/GitPackages/wtsGMT/data/depthcontour_200500.prn",
  basename = "SurveyMaps", cleanup = TRUE, verbosity = 1)
}
\arguments{
\item{tbl_cpue}{: dataframe from call to \code{\link{calcCPUE.ByHaul} or \code{\link{calcCPUE.ByStation}}} (or crab cpue filename, or NULL)}

\item{years}{: vector of survey years for which to produce maps}

\item{ztype}{: the data type (numCPUE, wgtCPUE to map)}

\item{scaleBy}{: factor to scale the ztype by BEFORE sending to plotMap.CSV}

\item{zlab}{: label for z (cpue) axes on map}

\item{zunits}{: label for z (cpue) units on maps}

\item{zscl}{: see Details}

\item{basename}{: base name for output files (year and factor level combinations will be added)}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
the max zscale, or a named list (names corresponding to individual factor combinations)
of max zscales that were used for plots. This value can be used to scale all maps
(if a single numeric value) or maps corresponding to a given factor combination (if
a named list) when running the function on the same cpue values a second time.
}
\description{
Function to map cpue by survey haul or station and other factors (e.g., sex) from cpue info.
}
\details{


Needs to filled in. \cr
Note: Multiple values at the same location will be summed prior to being mapped.
\cr
Note: zscl can be NULL, a numeric value, or a named list with numeric values as
list elements. If NULL, then the max z value for the data plotted on an individual
map will be used to scale that map. If a numeric value, then it is used as the
z-axis scale for all maps produced. If a named list, the element corresponding to
each combination of factors will be used as the z-axis scale for that set of maps.
}

