% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResampledACDs.R
\name{plotResampledACDs}
\alias{plotResampledACDs}
\title{Plot resampled aggregated catch data}
\usage{
plotResampledACDs(
  dfr,
  col = c("totBIOMASS", "totABUNDANCE"),
  scales = "free_y"
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe with resampled aggregated catch data
}}

\item{col}{\itemize{
\item column to plot
}}

\item{scales}{\itemize{
\item 'scales' parameter passed to facet_grid ("fixed" or "free_y")
}}
}
\description{
Function to plot resampled aggregated catch data.
}
\details{
dfr should be an output from \code{\link{calcResampledACDs}}.
}
