% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSizeComps.ByStratum.R
\name{plotSizeComps.ByStratum}
\alias{plotSizeComps.ByStratum}
\title{Plot size comps from survey data as bars.}
\usage{
plotSizeComps.ByStratum(zcs, facs = "", dropLevels = NULL,
  multipliers = NULL, var = c("ABUNDANCE", "BIOMASS"), rng = NULL,
  ggtheme = theme_grey(), ncol = 2, nrow = 5, showPlots = TRUE)
}
\arguments{
\item{zcs}{- dataframe from call to one of the calcSizeComps... functions}

\item{facs}{- factors to plot by}

\item{dropLevels}{- list by factor of factor levels to drop from plots}

\item{multipliers}{- list by factor of multipliers for factor levels}

\item{var}{- variable type to plot (abundance or biomass)}

\item{rng}{- y-axis range (calculated internally if NULL)}

\item{ggtheme}{- ggplot2 theme}

\item{ncol}{- number of columns of plots per page}

\item{nrow}{- number of rows of plots per page}

\item{showPlots}{- flag to show plots immediately}
}
\value{
list of ggplot2 plot objects
}
\description{
Function to plot size comps from survey data as bars.
}
\details{
The time series for each factor level combination is plotted
separately. Distinct levels of each factor can be dropped from the
final plot by seting dropLevels to a list with names corresponding to 
factor columns and values being vectors of factor levels to drop.

'multipliers' is a list of lists, with each sublist has two elements, 'factors' and 'value'.
'factors' should be a list of factor=level pairs. 'value' should be the value to multiply by. 

One plot is created for each distinct level of 'STRATUM'.
}

