% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledSizeComps.calcEffN.R
\name{resampledSizeComps.calcEffN}
\alias{resampledSizeComps.calcEffN}
\title{Calculate and plot effective N from resampled size compositions}
\usage{
resampledSizeComps.calcEffN(dfr, byFacs = "")
}
\arguments{
\item{dfr}{: dataframe with resampled size compositions}

\item{byFacs}{: vector of column names for factors other than YEAR and STRATUM}
}
\value{
a dataframe with the number of measured crab, the mean effective N's, the harmonic mean effective N's,
and the number of non-zero stations, by stratum, year, and factors in \code{byFacs}.
}
\description{
Function to calculate and plot effective N from resampled size compositions.
}
\details{
dfr should be a dataframe output from \link{resampledSizeComps.calc}.
}
