% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledSizeComps.plot.R
\name{resampledSizeComps.plot}
\alias{resampledSizeComps.plot}
\title{Plot resampled size compositions}
\usage{
resampledSizeComps.plot(dfr, scales = "free_y", dropLevels = list(SEX =
  c("HERMAPHRODITE", "HERMAPHRODITIC", "UNDETERMINED")),
  convertImmOS = TRUE)
}
\arguments{
\item{dfr}{- dataframe with resampled size compositions}

\item{scales}{- 'scales' parameter passed to facet_grid ("fixed" or "free_y")}

\item{dropLevels}{- list of factor levels to remove from dfr}

\item{convertImmOS}{- flag to convert immature, old shell crab to immature, new shell crab}
}
\description{
Function to plot resampled size compositions.
}
\details{
dfr should be an output from \code{\link{resampledSizeComps.calc}}.
}
