% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledSizeComps.plot.R
\name{resampledSizeComps.plot}
\alias{resampledSizeComps.plot}
\title{Plot resampled size compositions}
\usage{
resampledSizeComps.plot(dfr, scales = "free_y")
}
\arguments{
\item{dfr}{\itemize{
\item dataframe with resampled size compositions
}}

\item{scales}{\itemize{
\item 'scales' parameter passed to facet_grid ("fixed" or "free_y")
}}
}
\description{
Function to plot resampled size compositions.
}
\details{
dfr should be an output from \link{resampledSizeComps.calc}.
}
