% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectIndivs.TrawlSurvey.R
\name{selectIndivs.TrawlSurvey}
\alias{selectIndivs.TrawlSurvey}
\title{Function to extract crab data on individuals w/ selected characteristics from NMFS trawl survey csv files.}
\usage{
selectIndivs.TrawlSurvey(tbl_hauls, tbl = NULL, col.Size = "WIDTH",
  sex = c("MALE", "FEMALE", "ALL"), shell_condition = c("NEW_SHELL",
  "OLD_SHELL", "ALL"), maturity = c("IMMATURE", "MATURE", "ALL"),
  calcMaleMaturity = FALSE, minSize = -Inf, maxSize = Inf,
  export = FALSE, out.csv = "SelectedIndivs.csv", out.dir = NULL,
  verbosity = 0)
}
\arguments{
\item{tbl_hauls}{: hauls table (dataframe) from call to selectHauls.TrawlSurvey(...) [required]}

\item{tbl}{: table (dataframe) of survey data (or csv filename or NULL)}

\item{col.Size}{: column name for size information}

\item{sex}{: one of 'MALE','FEMALE' or 'ALL' for narrowing selection of individuals}

\item{shell_condition}{: one of 'NEW_SHELL','OLD_SHELL' or 'ALL' for narrowing selection of individuals}

\item{maturity}{: one of 'IMMATURE','MATURE' or 'ALL' for narrowing selection of individuals}

\item{calcMaleMaturity}{: flag (T/F) to calculate pr(mature|size) for males based on an ogive}

\item{minSize}{: minimum size (width) of individuals to select}

\item{maxSize}{: maximum size (width) of individuals to select}

\item{export}{- boolean flag to export results to csv file}

\item{out.csv}{- name of output csv file                    (ignored if NULL)}

\item{out.dir}{- base path for output csv file              (set to folder of input csv file or current working directory)}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}
}
\value{
dataframe (see Details for coulmn names)
}
\description{
Function to extract crab data on individuals w/ selected characteristics from NMFS trawl survey csv files.
}
\details{
If neither tbl or in.csv is given, the user will be prompted for a csv file via a file dialog box.\cr
Returned dataframe will have columns:
\itemize{\item {HAULJOIN}
         \item {numIndivs}
         \item {SEX}
         \item {SHELL_CONDITION}
         \item {MATURITY}
         \item {SEX_CODE}
         \item {SHELL_CONDITION_CODE}
         \item {SIZE}
         \item {EGG_COLOR}
         \item {EGG_CONDITION}
         \item {CLUTCH_SIZE}
         \item {CHERLA_HEIGHT}
         \item {SAMPLING_FACTOR}
         \item {WEIGHT}
         \item {CALCULATED_WEIGHT}
        } \cr
Notes:
\itemize{\item Weights are in grams.}
}
