% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/selectStrata.TrawlSurvey.R
\name{selectStrata.TrawlSurvey}
\alias{selectStrata.TrawlSurvey}
\title{Function to extract crab station data and strata information from a NMFS trawl survey strata file.}
\usage{
selectStrata.TrawlSurvey(tbl = NULL, species = "BTC",
  useOrigStrata = FALSE, export = FALSE,
  out.csv = paste("SelectedStations", species, "csv", sep = "."),
  out.dir = NULL, verbosity = 1)
}
\arguments{
\item{tbl}{: table (dataframe) of survey station/strata data (or name of survey station/strata csv file, or NULL)}

\item{species}{: code ('BKC','BTC','RKC','OTC') indicating species}

\item{export}{: boolean flag to export results to csv file}

\item{verbosity}{: integer flag indicating level of printed output (0=off,1=minimal,2=full)}

\item{useOrigStrata:}{boolean to use Bob Foy's "original" strata, rather than the revised strata}
}
\value{
a dataframe with strata/stations info.
}
\description{
Same as selectStations.TrawlSurvey(...)
}
\details{


If tbl is NULL, the user will be prompted for a csv file via a file dialog box.
Returned dataframe has columns:
\itemize{\item {YEAR}
        \item {STRATUM}
        \item {STRATUM_CODE}
        \item {STRATUM_AREA}
        \item {GIS_STATION}
        \item {STATION_LONGITUDE}
        \item {STATION_LATITUDE}
        }
}

