% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subForTCSAM.R
\name{subForTCSAM}
\alias{subForTCSAM}
\title{Check abnd make substitutions for "undetermined", "unknown", or "all"}
\usage{
subForTCSAM(x, str)
}
\arguments{
\item{x}{- character string to check if equal to "undetermined", "unknown", or "all"}

\item{str}{- character string to substitute}
}
\value{
character string with potential susbstitutions made
}
\description{
Function to check and make substitutions for "undetermined", "unknown", or "all".
}
\details{
If \code{x} is"undetermined", "unknown", or "all", the alternative character string
\code{str} is returned, otherwise the original is returned
}
